Add initial character levelling changes
