+++ src/endpoints/character/CharacterCreationView.tsx
@@ -1,7 +1,6 @@
 ﻿import { useCallback, useEffect, useMemo, useState, type Dispatch, type SetStateAction } from 'react';
 
 import {
-  applyLevellingChoices,
   fetchCultureTypes,
   fetchCultures,
   fetchLanguages,
@@ -61,7 +60,6 @@ type CharacterStep =
   | 'physique'
   | 'hobby'
   | 'background'
-  | 'apprenticeship'
   | 'summary';
 
 const STEP_ORDER: CharacterStep[] = [
@@ -71,7 +69,6 @@ const STEP_ORDER: CharacterStep[] = [
   'physique',
   'hobby',
   'background',
-  'apprenticeship',
   'summary',
 ];
 
@@ -82,8 +79,7 @@ const STEP_LABELS: Record<CharacterStep, string> = {
   physique: '4. Physique',
   hobby: '5. Hobby Ranks',
   background: '6. Background Options',
-  apprenticeship: '7. Apprenticeship Skills',
-  summary: '8. Summary',
+  summary: '7. Summary',
 };
 
 const OWN_REALM_OPEN_LISTS_CATEGORY_ID = 'SKILLCATEGORY_SPELLS_OWN_REALM_OPEN_LISTS';
@@ -97,7 +93,6 @@ type StepErrors = {
   physique?: string | undefined;
   hobby?: string | undefined;
   background?: string | undefined;
-  apprenticeship?: string | undefined;
   summary?: string | undefined;
 };
 
@@ -157,42 +152,6 @@ type BackgroundOptionState = {
   categoryBonusIds: string[];
   specialItemsPoints: number;
 };
-
-type LevellingSkillPurchase = {
-  id: string;
-  subcategory: string;
-  purchases: number;
-};
-
-type LevellingCategoryPurchase = {
-  id: string;
-  purchases: number;
-};
-
-type LevellingSpellListPurchase = {
-  id: string;
-  purchases: number;
-};
-
-type TpSkillAllocation = { id: string; subcategory: string; ranks: number };
-type TpSpellListAllocation = { id: string; ranks: number };
-type TpGroupCategoryAndSkillChoice = { categoryId: string; skillId: string; subcategory: string };
-type TpLanguageAllocation = { languageId: string; spoken: number; written: number; somatic: number };
-
-type TpResolution = {
-  tpId: string;
-  statGainChoices: (Stat | '')[];
-  skillRankChoices: TpSkillAllocation[][];
-  categoryMultiSkillChoices: TpSkillAllocation[][];
-  groupMultiSkillChoices: TpSkillAllocation[][];
-  groupCategoryAndSkillChoices: TpGroupCategoryAndSkillChoice[];
-  spellListChoices: TpSpellListAllocation[][];
-  spellListCategoryChoices: TpSpellListAllocation[][];
-  lifestyleCategorySkillChoices: string[][];
-  languageChoices: TpLanguageAllocation[][];
-};
-
-type LevellingSubstep = 'selecting' | 'resolving' | 'purchasing';
 
 type StatRoll = {
   slot: number;
@@ -332,20 +291,9 @@ function getSelectedBackgroundOptionsPayload(state: BackgroundOptionState): stri
   return out;
 }
 
-const STAT_GAIN_DP_COST = 8;
-
 function parseCostString(cost: string): number[] {
   if (!cost) return [];
   return cost.split(':').map(Number).filter((n) => !isNaN(n) && n >= 0);
-}
-
-function getMaxPurchases(costElements: number[]): number {
-  return costElements.length;
-}
-
-function getSkillMaxPurchases(costElements: number[], devType: SkillDevelopmentType | undefined): number {
-  if (devType === 'Restricted') return Math.ceil(costElements.length / 2);
-  return costElements.length;
 }
 
 function getSkillRanksPerPurchase(devType: SkillDevelopmentType | undefined): number {
@@ -356,247 +304,6 @@ function getSkillRanksPerPurchase(devType: SkillDevelopmentType | undefined): nu
     case 'Standard':
     default: return 1;
   }
-}
-
-function getSkillPurchaseTotalCost(costElements: number[], devType: SkillDevelopmentType | undefined, purchases: number): number {
-  if (devType === 'Restricted') {
-    let total = 0;
-    for (let r = 0; r < purchases; r++) {
-      const i1 = r * 2;
-      const i2 = r * 2 + 1;
-      if (i2 < costElements.length) {
-        total += (costElements[i1] ?? 0) + (costElements[i2] ?? 0);
-      } else if (i1 < costElements.length) {
-        total += (costElements[i1] ?? 0) * 2;
-      }
-    }
-    return total;
-  }
-  return costElements.slice(0, purchases).reduce((s, c) => s + c, 0);
-}
-
-function getCategoryOrSpellListPurchaseTotalCost(costElements: number[], purchases: number): number {
-  return costElements.slice(0, purchases).reduce((s, c) => s + c, 0);
-}
-
-function tpHasChoices(tp: TrainingPackage): boolean {
-  return (
-    (tp.statGainChoices?.numChoices ?? 0) > 0
-    || (tp.skillRankChoices ?? []).some((c) => c.numChoices > 0)
-    || (tp.categoryMultiSkillRankChoices ?? []).some((c) => c.numChoices > 0)
-    || (tp.groupMultiSkillRankChoices ?? []).some((c) => c.numChoices > 0)
-    || (tp.groupCategoryAndSkillRankChoices ?? []).length > 0
-    || (tp.spellListRanks ?? []).some((r) => r.numChoices > 0)
-    || (tp.spellListCategoryRankChoices ?? []).some((c) => c.numChoices > 0)
-    || (tp.lifestyleCategorySkillChoices ?? []).some((c) => c.numChoices > 0)
-    || (tp.languageChoices ?? []).some((c) => c.numChoices > 0)
-  );
-}
-
-function createEmptyTpResolution(tp: TrainingPackage): TpResolution {
-  return {
-    tpId: tp.id,
-    statGainChoices: Array.from({ length: tp.statGainChoices?.numChoices ?? 0 }, () => '' as Stat | ''),
-    skillRankChoices: (tp.skillRankChoices ?? []).map(() => []),
-    categoryMultiSkillChoices: (tp.categoryMultiSkillRankChoices ?? []).map(() => []),
-    groupMultiSkillChoices: (tp.groupMultiSkillRankChoices ?? []).map(() => []),
-    groupCategoryAndSkillChoices: (tp.groupCategoryAndSkillRankChoices ?? []).map(() => ({
-      categoryId: '',
-      skillId: '',
-      subcategory: '',
-    })),
-    spellListChoices: (tp.spellListRanks ?? [])
-      .filter((r) => r.numChoices > 0)
-      .map(() => []),
-    spellListCategoryChoices: (tp.spellListCategoryRankChoices ?? []).map(() => []),
-    lifestyleCategorySkillChoices: (tp.lifestyleCategorySkillChoices ?? []).map((c) =>
-      Array.from({ length: c.numChoices }, () => ''),
-    ),
-    languageChoices: (tp.languageChoices ?? []).map(() => []),
-  };
-}
-
-function validateTpResolution(tp: TrainingPackage, resolution: TpResolution, excludedStats: ReadonlySet<Stat> = new Set()): string | undefined {
-  if (tp.statGainChoices && tp.statGainChoices.numChoices > 0) {
-    for (let i = 0; i < tp.statGainChoices.numChoices; i++) {
-      if (!resolution.statGainChoices[i]) return `${tp.name}: select stat for gain choice ${i + 1}.`;
-      if (excludedStats.has(resolution.statGainChoices[i] as Stat)) {
-        return `${tp.name}: stat gain choice ${i + 1} is already claimed by a realm stat gain.`;
-      }
-    }
-    const chosen = resolution.statGainChoices.filter(Boolean);
-    if (new Set(chosen).size < chosen.length) return `${tp.name}: stat gain choices must be unique.`;
-  }
-
-  for (let gi = 0; gi < (tp.skillRankChoices ?? []).length; gi++) {
-    const choiceDef = (tp.skillRankChoices ?? [])[gi];
-    if (!choiceDef || choiceDef.numChoices <= 0) continue;
-    const allocs = resolution.skillRankChoices[gi] ?? [];
-    const totalRanks = allocs.reduce((s, a) => s + a.ranks, 0);
-    if (totalRanks !== choiceDef.value) {
-      const remaining = choiceDef.value - totalRanks;
-      return `${tp.name}: skill rank choice ${gi + 1}: ${remaining > 0 ? `${remaining} rank${remaining !== 1 ? 's' : ''} still to allocate` : 'too many ranks allocated'}.`;
-    }
-    if (allocs.length > choiceDef.numChoices) return `${tp.name}: skill rank choice ${gi + 1}: at most ${choiceDef.numChoices} skill${choiceDef.numChoices !== 1 ? 's' : ''} allowed.`;
-    for (const alloc of allocs) {
-      if (!alloc.id) return `${tp.name}: skill rank choice ${gi + 1}: all allocations must have a skill selected.`;
-      if (alloc.ranks < 1) return `${tp.name}: skill rank choice ${gi + 1}: each allocation must have at least 1 rank.`;
-    }
-    const ids = allocs.map((a) => a.id);
-    if (new Set(ids).size < ids.length) return `${tp.name}: skill rank choice ${gi + 1}: duplicate skills selected.`;
-  }
-
-  for (let gi = 0; gi < (tp.categoryMultiSkillRankChoices ?? []).length; gi++) {
-    const choiceDef = (tp.categoryMultiSkillRankChoices ?? [])[gi];
-    if (!choiceDef || choiceDef.numChoices <= 0) continue;
-    const allocs = resolution.categoryMultiSkillChoices[gi] ?? [];
-    const totalRanks = allocs.reduce((s, a) => s + a.ranks, 0);
-    if (totalRanks !== choiceDef.value) {
-      const remaining = choiceDef.value - totalRanks;
-      return `${tp.name}: category skill choice ${gi + 1}: ${remaining > 0 ? `${remaining} rank${remaining !== 1 ? 's' : ''} still to allocate` : 'too many ranks allocated'}.`;
-    }
-    if (allocs.length > choiceDef.numChoices) return `${tp.name}: category skill choice ${gi + 1}: at most ${choiceDef.numChoices} skill${choiceDef.numChoices !== 1 ? 's' : ''} allowed.`;
-    for (const alloc of allocs) {
-      if (!alloc.id) return `${tp.name}: category skill choice ${gi + 1}: all allocations must have a skill selected.`;
-      if (alloc.ranks < 1) return `${tp.name}: category skill choice ${gi + 1}: each allocation must have at least 1 rank.`;
-    }
-    const ids = allocs.map((a) => a.id);
-    if (new Set(ids).size < ids.length) return `${tp.name}: category skill choice ${gi + 1}: duplicate skills selected.`;
-  }
-
-  for (let gi = 0; gi < (tp.groupMultiSkillRankChoices ?? []).length; gi++) {
-    const choiceDef = (tp.groupMultiSkillRankChoices ?? [])[gi];
-    if (!choiceDef || choiceDef.numChoices <= 0) continue;
-    const allocs = resolution.groupMultiSkillChoices[gi] ?? [];
-    const totalRanks = allocs.reduce((s, a) => s + a.ranks, 0);
-    if (totalRanks !== choiceDef.value) {
-      const remaining = choiceDef.value - totalRanks;
-      return `${tp.name}: group skill choice ${gi + 1}: ${remaining > 0 ? `${remaining} rank${remaining !== 1 ? 's' : ''} still to allocate` : 'too many ranks allocated'}.`;
-    }
-    if (allocs.length > choiceDef.numChoices) return `${tp.name}: group skill choice ${gi + 1}: at most ${choiceDef.numChoices} skill${choiceDef.numChoices !== 1 ? 's' : ''} allowed.`;
-    for (const alloc of allocs) {
-      if (!alloc.id) return `${tp.name}: group skill choice ${gi + 1}: all allocations must have a skill selected.`;
-      if (alloc.ranks < 1) return `${tp.name}: group skill choice ${gi + 1}: each allocation must have at least 1 rank.`;
-    }
-    const ids = allocs.map((a) => a.id);
-    if (new Set(ids).size < ids.length) return `${tp.name}: group skill choice ${gi + 1}: duplicate skills selected.`;
-  }
-
-  for (let gi = 0; gi < (tp.groupCategoryAndSkillRankChoices ?? []).length; gi++) {
-    const slot = resolution.groupCategoryAndSkillChoices[gi];
-    if (!slot || !slot.categoryId) return `${tp.name}: group category & skill choice ${gi + 1}: select a category.`;
-    if (!slot.skillId) return `${tp.name}: group category & skill choice ${gi + 1}: select a skill.`;
-  }
-
-  const choiceableSlRanks = (tp.spellListRanks ?? []).filter((r) => r.numChoices > 0);
-  for (let gi = 0; gi < choiceableSlRanks.length; gi++) {
-    const choiceDef = choiceableSlRanks[gi];
-    if (!choiceDef) continue;
-    const allocs = resolution.spellListChoices[gi] ?? [];
-    const totalRanks = allocs.reduce((s, a) => s + a.ranks, 0);
-    if (totalRanks !== choiceDef.value) {
-      const remaining = choiceDef.value - totalRanks;
-      return `${tp.name}: spell list choice ${gi + 1}: ${remaining > 0 ? `${remaining} rank${remaining !== 1 ? 's' : ''} still to allocate` : 'too many ranks allocated'}.`;
-    }
-    if (allocs.length > choiceDef.numChoices) return `${tp.name}: spell list choice ${gi + 1}: at most ${choiceDef.numChoices} spell list${choiceDef.numChoices !== 1 ? 's' : ''} allowed.`;
-    for (const alloc of allocs) {
-      if (!alloc.id) return `${tp.name}: spell list choice ${gi + 1}: all allocations must have a spell list selected.`;
-      if (alloc.ranks < 1) return `${tp.name}: spell list choice ${gi + 1}: each allocation must have at least 1 rank.`;
-    }
-    const ids = allocs.map((a) => a.id);
-    if (new Set(ids).size < ids.length) return `${tp.name}: spell list choice ${gi + 1}: duplicate spell lists selected.`;
-  }
-
-  for (let gi = 0; gi < (tp.spellListCategoryRankChoices ?? []).length; gi++) {
-    const choiceDef = (tp.spellListCategoryRankChoices ?? [])[gi];
-    if (!choiceDef || choiceDef.numChoices <= 0) continue;
-    const allocs = resolution.spellListCategoryChoices[gi] ?? [];
-    const totalRanks = allocs.reduce((s, a) => s + a.ranks, 0);
-    if (totalRanks !== choiceDef.value) {
-      const remaining = choiceDef.value - totalRanks;
-      return `${tp.name}: spell list category choice ${gi + 1}: ${remaining > 0 ? `${remaining} rank${remaining !== 1 ? 's' : ''} still to allocate` : 'too many ranks allocated'}.`;
-    }
-    if (allocs.length > choiceDef.numChoices) return `${tp.name}: spell list category choice ${gi + 1}: at most ${choiceDef.numChoices} spell list${choiceDef.numChoices !== 1 ? 's' : ''} allowed.`;
-    for (const alloc of allocs) {
-      if (!alloc.id) return `${tp.name}: spell list category choice ${gi + 1}: all allocations must have a spell list selected.`;
-      if (alloc.ranks < 1) return `${tp.name}: spell list category choice ${gi + 1}: each allocation must have at least 1 rank.`;
-    }
-    const ids = allocs.map((a) => a.id);
-    if (new Set(ids).size < ids.length) return `${tp.name}: spell list category choice ${gi + 1}: duplicate spell lists selected.`;
-  }
-
-  for (let gi = 0; gi < (tp.lifestyleCategorySkillChoices ?? []).length; gi++) {
-    const choiceDef = (tp.lifestyleCategorySkillChoices ?? [])[gi];
-    if (!choiceDef || choiceDef.numChoices <= 0) continue;
-    const chosen = resolution.lifestyleCategorySkillChoices[gi] ?? [];
-    for (let si = 0; si < choiceDef.numChoices; si++) {
-      if (!chosen[si]) return `${tp.name}: lifestyle skill choice ${gi + 1} slot ${si + 1} required.`;
-    }
-    const nonEmpty = chosen.filter(Boolean);
-    if (new Set(nonEmpty).size < nonEmpty.length) return `${tp.name}: lifestyle skill choice ${gi + 1}: duplicate skills selected.`;
-  }
-
-  for (let gi = 0; gi < (tp.languageChoices ?? []).length; gi++) {
-    const choiceDef = (tp.languageChoices ?? [])[gi];
-    if (!choiceDef || choiceDef.numChoices <= 0) continue;
-    const allocs = resolution.languageChoices[gi] ?? [];
-    if (allocs.length === 0) return `${tp.name}: language choice ${gi + 1}: at least one language required.`;
-    if (allocs.length > choiceDef.numChoices) return `${tp.name}: language choice ${gi + 1}: at most ${choiceDef.numChoices} language${choiceDef.numChoices !== 1 ? 's' : ''} allowed.`;
-    for (const alloc of allocs) {
-      if (!alloc.languageId) return `${tp.name}: language choice ${gi + 1}: all entries must have a language selected.`;
-    }
-    const langIds = allocs.map((a) => a.languageId).filter(Boolean);
-    if (new Set(langIds).size < langIds.length) return `${tp.name}: language choice ${gi + 1}: duplicate languages selected.`;
-    const totalRanks = allocs.reduce((s, a) => s + a.spoken + a.written + a.somatic, 0);
-    if (totalRanks !== choiceDef.value) {
-      const remaining = choiceDef.value - totalRanks;
-      return `${tp.name}: language choice ${gi + 1}: ${remaining > 0 ? `${remaining} rank${remaining !== 1 ? 's' : ''} still to allocate` : `${Math.abs(remaining)} rank${Math.abs(remaining) !== 1 ? 's' : ''} over`}.`;
-    }
-  }
-  return undefined;
-}
-
-function getSkillMaxDpPurchases(
-  costElements: number[],
-  devType: SkillDevelopmentType | undefined,
-  tpRanks: number,
-): number {
-  if (devType === 'Restricted') {
-    return Math.max(0, Math.floor((costElements.length - tpRanks * 2) / 2));
-  }
-  const ranksPerPurchase = getSkillRanksPerPurchase(devType);
-  const usedPurchases = Math.ceil(tpRanks / (ranksPerPurchase || 1));
-  return Math.max(0, costElements.length - usedPurchases);
-}
-
-function getSkillDpCostWithTpOffset(
-  costElements: number[],
-  devType: SkillDevelopmentType | undefined,
-  purchases: number,
-  tpRanks: number,
-): number {
-  if (purchases <= 0) return 0;
-  if (devType === 'Restricted') {
-    let total = 0;
-    for (let r = 0; r < purchases; r++) {
-      const i1 = (tpRanks + r) * 2;
-      const i2 = (tpRanks + r) * 2 + 1;
-      total += (costElements[i1] ?? 0) + (costElements[i2] ?? 0);
-    }
-    return total;
-  }
-  const ranksPerPurchase = getSkillRanksPerPurchase(devType);
-  const purchaseOffset = Math.ceil(tpRanks / (ranksPerPurchase || 1));
-  return costElements.slice(purchaseOffset, purchaseOffset + purchases).reduce((s, c) => s + c, 0);
-}
-
-function getCategoryMaxDpPurchases(costElements: number[], tpRanks: number): number {
-  return Math.max(0, costElements.length - tpRanks);
-}
-
-function getCategoryDpCostWithTpOffset(costElements: number[], purchases: number, tpRanks: number): number {
-  if (purchases <= 0) return 0;
-  return costElements.slice(tpRanks, tpRanks + purchases).reduce((s, c) => s + c, 0);
 }
 
 function getBuildLabel(modifier: number): string {
@@ -683,24 +390,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
   const [backgroundCategoryBonusIds, setBackgroundCategoryBonusIds] = useState<string[]>([]);
   const [backgroundSpecialItemsPoints, setBackgroundSpecialItemsPoints] = useState(0);
 
-  const [levellingTrainingPackageIds, setLevellingTrainingPackageIds] = useState<string[]>([]);
-  const [levellingStatGains, setLevellingStatGains] = useState<Stat[]>([]);
-  const [levellingSkillPurchases, setLevellingSkillPurchases] = useState<LevellingSkillPurchase[]>([]);
-  const [levellingSkillCategoryFilter, setLevellingSkillCategoryFilter] = useState('');
-  const [levellingSkillPendingId, setLevellingSkillPendingId] = useState('');
-  const [levellingSkillPendingSubcategory, setLevellingSkillPendingSubcategory] = useState('');
-  const [levellingCategoryPurchases, setLevellingCategoryPurchases] = useState<LevellingCategoryPurchase[]>([]);
-  const [levellingSpellListPurchases, setLevellingSpellListPurchases] = useState<LevellingSpellListPurchase[]>([]);
-  const [levellingSelectedSpellCategory, setLevellingSelectedSpellCategory] = useState('');
-  const [levellingAddingSpellList, setLevellingAddingSpellList] = useState(false);
-  const [levellingSubstep, setLevellingSubstep] = useState<LevellingSubstep>('selecting');
-  const [levellingResolvingTpIndex, setLevellingResolvingTpIndex] = useState(0);
-  const [tpResolutions, setTpResolutions] = useState<TpResolution[]>([]);
-  const [preLevellingRanks, setPreLevellingRanks] = useState<{
-    skillRanks: CharacterBuilder['skillRanks'];
-    categoryRanks: CharacterBuilder['categoryRanks'];
-    spellListRanks: CharacterBuilder['spellListRanks'];
-  }>({ skillRanks: [], categoryRanks: [], spellListRanks: [] });
   const [characterBuilder, setCharacterBuilder] = useState<CharacterBuilder>(() => createEmptyCharacterBuilder());
   const [savingPrimaryDefinition, setSavingPrimaryDefinition] = useState(false);
   const [savingInitialChoices, setSavingInitialChoices] = useState(false);
@@ -1283,21 +972,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
     return [...preferred, ...rest];
   }, [trainingPackages, raceId, culture]);
 
-  const selectedLevellingTrainingPackages = useMemo(
-    () => levellingTrainingPackageIds
-      .map((id) => availableTrainingPackages.find((tp) => tp.id === id))
-      .filter((tp): tp is TrainingPackage => tp !== undefined),
-    [availableTrainingPackages, levellingTrainingPackageIds],
-  );
-
-  const tpCostMap = useMemo(() => {
-    const map = new Map<string, number>();
-    for (const row of characterBuilder.trainingPackageCosts) {
-      map.set(row.id, row.value);
-    }
-    return map;
-  }, [characterBuilder.trainingPackageCosts]);
-
   const categoryCostMap = useMemo(() => {
     const map = new Map<string, number[]>();
     for (const row of characterBuilder.categoryCosts) {
@@ -1319,283 +993,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
     for (const s of skills) map.set(s.id, s.category);
     return map;
   }, [skills]);
-
-  const levellingSelectedLifestylePackageId = useMemo(() => {
-    return levellingTrainingPackageIds.find((tpId) => {
-      const tp = trainingPackages.find((t) => t.id === tpId);
-      return tp?.lifestyle;
-    }) ?? null;
-  }, [levellingTrainingPackageIds, trainingPackages]);
-
-  const levellingStatGainsUnavailable = useMemo((): Set<Stat> => {
-    const claimed = new Set<Stat>();
-    for (const tp of selectedLevellingTrainingPackages) {
-      for (const stat of tp.statGains ?? []) {
-        claimed.add(stat);
-      }
-      if (tp.realmStatGain) {
-        for (const realm of characterBuilder.magicalRealms) {
-          const stat = getStatForRealm(realm);
-          if (stat) claimed.add(stat);
-        }
-      }
-      const res = tpResolutions.find((r) => r.tpId === tp.id);
-      if (res) {
-        for (const stat of res.statGainChoices) {
-          if (stat) claimed.add(stat);
-        }
-      }
-    }
-    return claimed;
-  }, [selectedLevellingTrainingPackages, tpResolutions, characterBuilder.magicalRealms]);
-
-  useEffect(() => {
-    if (levellingStatGainsUnavailable.size === 0) return;
-    setLevellingStatGains((prev) => prev.filter((s) => !levellingStatGainsUnavailable.has(s)));
-  }, [levellingStatGainsUnavailable]);
-
-  const tpGrantedSkillRankCounts = useMemo(() => {
-    const map = new Map<string, number>();
-    for (const tp of selectedLevellingTrainingPackages) {
-      for (const rank of tp.skillRanks ?? []) {
-        map.set(rank.id, (map.get(rank.id) ?? 0) + rank.value);
-      }
-      const res = tpResolutions.find((r) => r.tpId === tp.id);
-      if (!res) continue;
-      for (let gi = 0; gi < (tp.skillRankChoices ?? []).length; gi++) {
-        for (const alloc of res.skillRankChoices[gi] ?? []) {
-          if (alloc.id) map.set(alloc.id, (map.get(alloc.id) ?? 0) + alloc.ranks);
-        }
-      }
-      for (let gi = 0; gi < (tp.categoryMultiSkillRankChoices ?? []).length; gi++) {
-        for (const alloc of res.categoryMultiSkillChoices[gi] ?? []) {
-          if (alloc.id) map.set(alloc.id, (map.get(alloc.id) ?? 0) + alloc.ranks);
-        }
-      }
-      for (let gi = 0; gi < (tp.groupMultiSkillRankChoices ?? []).length; gi++) {
-        for (const alloc of res.groupMultiSkillChoices[gi] ?? []) {
-          if (alloc.id) map.set(alloc.id, (map.get(alloc.id) ?? 0) + alloc.ranks);
-        }
-      }
-      for (let gi = 0; gi < (tp.groupCategoryAndSkillRankChoices ?? []).length; gi++) {
-        const choiceDef = (tp.groupCategoryAndSkillRankChoices ?? [])[gi];
-        const slot = res.groupCategoryAndSkillChoices[gi];
-        if (choiceDef && slot?.skillId) {
-          map.set(slot.skillId, (map.get(slot.skillId) ?? 0) + choiceDef.value);
-        }
-      }
-    }
-    return map;
-  }, [selectedLevellingTrainingPackages, tpResolutions]);
-
-  const tpGrantedCategoryRankCounts = useMemo(() => {
-    const map = new Map<string, number>();
-    for (const tp of selectedLevellingTrainingPackages) {
-      for (const rank of tp.categoryRanks ?? []) {
-        map.set(rank.id, (map.get(rank.id) ?? 0) + rank.value);
-      }
-      const res = tpResolutions.find((r) => r.tpId === tp.id);
-      if (!res) continue;
-      for (let gi = 0; gi < (tp.groupCategoryAndSkillRankChoices ?? []).length; gi++) {
-        const choiceDef = (tp.groupCategoryAndSkillRankChoices ?? [])[gi];
-        const slot = res.groupCategoryAndSkillChoices[gi];
-        if (choiceDef && slot?.categoryId) {
-          map.set(slot.categoryId, (map.get(slot.categoryId) ?? 0) + choiceDef.value);
-        }
-      }
-    }
-    return map;
-  }, [selectedLevellingTrainingPackages, tpResolutions]);
-
-  const tpGrantedSpellListRankCounts = useMemo(() => {
-    const map = new Map<string, number>();
-    for (const tp of selectedLevellingTrainingPackages) {
-      const res = tpResolutions.find((r) => r.tpId === tp.id);
-      if (!res) continue;
-      const choiceableSlRanks = (tp.spellListRanks ?? []).filter((r) => r.numChoices > 0);
-      for (let gi = 0; gi < choiceableSlRanks.length; gi++) {
-        for (const alloc of res.spellListChoices[gi] ?? []) {
-          if (alloc.id) map.set(alloc.id, (map.get(alloc.id) ?? 0) + alloc.ranks);
-        }
-      }
-      for (let gi = 0; gi < (tp.spellListCategoryRankChoices ?? []).length; gi++) {
-        for (const alloc of res.spellListCategoryChoices[gi] ?? []) {
-          if (alloc.id) map.set(alloc.id, (map.get(alloc.id) ?? 0) + alloc.ranks);
-        }
-      }
-    }
-    return map;
-  }, [selectedLevellingTrainingPackages, tpResolutions]);
-
-  const tpsRequiringResolution = useMemo(
-    () => selectedLevellingTrainingPackages.filter(tpHasChoices),
-    [selectedLevellingTrainingPackages],
-  );
-
-  const levellingTrainingPackageDpCost = useMemo(() => {
-    return levellingTrainingPackageIds.reduce((total, tpId) => {
-      return total + (tpCostMap.get(tpId) ?? 0);
-    }, 0);
-  }, [levellingTrainingPackageIds, tpCostMap]);
-
-  const levellingStatGainDpCost = levellingStatGains.length * STAT_GAIN_DP_COST;
-
-  const levellingSkillDpCost = useMemo(() => {
-    return levellingSkillPurchases.reduce((total, p) => {
-      const categoryId = skillCategoryMap.get(p.id);
-      if (!categoryId) return total;
-      const costElements = categoryCostMap.get(categoryId) ?? [];
-      const devType = skillDevTypeMap.get(p.id);
-      const tpRanks = tpGrantedSkillRankCounts.get(p.id) ?? 0;
-      return total + getSkillDpCostWithTpOffset(costElements, devType, p.purchases, tpRanks);
-    }, 0);
-  }, [levellingSkillPurchases, categoryCostMap, skillDevTypeMap, skillCategoryMap, tpGrantedSkillRankCounts]);
-
-  const levellingCategoryDpCost = useMemo(() => {
-    return levellingCategoryPurchases.reduce((total, p) => {
-      const costElements = categoryCostMap.get(p.id) ?? [];
-      const tpRanks = tpGrantedCategoryRankCounts.get(p.id) ?? 0;
-      return total + getCategoryDpCostWithTpOffset(costElements, p.purchases, tpRanks);
-    }, 0);
-  }, [levellingCategoryPurchases, categoryCostMap, tpGrantedCategoryRankCounts]);
-
-  const levellingSpellListDpCost = useMemo(() => {
-    return levellingSpellListPurchases.reduce((total, p) => {
-      const catEntry = characterBuilder.categorySpellLists.find((c) => c.spellLists.includes(p.id));
-      if (!catEntry) return total;
-      const costElements = categoryCostMap.get(catEntry.category) ?? [];
-      const tpRanks = tpGrantedSpellListRankCounts.get(p.id) ?? 0;
-      return total + getCategoryDpCostWithTpOffset(costElements, p.purchases, tpRanks);
-    }, 0);
-  }, [levellingSpellListPurchases, categoryCostMap, characterBuilder.categorySpellLists, tpGrantedSpellListRankCounts]);
-
-  const levellingTotalDpSpent = levellingTrainingPackageDpCost + levellingStatGainDpCost + levellingSkillDpCost + levellingCategoryDpCost + levellingSpellListDpCost;
-  const levellingDpRemaining = characterBuilder.developmentPoints - levellingTotalDpSpent;
-
-  const levellingTrainingPackageOptions = useMemo(() => {
-    const selectedSet = new Set(levellingTrainingPackageIds);
-    return availableTrainingPackages
-      .filter((tp) => {
-        if (selectedSet.has(tp.id)) return false;
-        if (tp.lifestyle && levellingSelectedLifestylePackageId !== null) return false;
-        const cost = tpCostMap.get(tp.id) ?? 0;
-        if (cost > levellingDpRemaining) return false;
-        return true;
-      })
-      .sort((a, b) => a.name.localeCompare(b.name))
-      .map((tp): RichSelectOption => {
-        const cost = tpCostMap.get(tp.id) ?? '?';
-        const secondary = tp.lifestyle ? `(Lifestyle) \u2014 ${cost} DP` : `${cost} DP`;
-        return {
-          value: tp.id,
-          label: <RichOptionLabel primary={tp.name} secondary={secondary} />,
-          searchText: `${tp.name} ${secondary}`,
-          title: showDescriptions && tp.description ? tp.description : undefined,
-        };
-      });
-  }, [availableTrainingPackages, levellingTrainingPackageIds, levellingSelectedLifestylePackageId, tpCostMap, levellingDpRemaining, showDescriptions]);
-
-  const levellingAvailableSkills = useMemo(() => {
-    const selectedSet = new Set(levellingSkillPurchases.map((p) => p.id));
-    return skills.filter((s) => {
-      if (selectedSet.has(s.id)) return false;
-      const costElements = categoryCostMap.get(s.category) ?? [];
-      const devType = skillDevTypeMap.get(s.id);
-      const tpRanks = tpGrantedSkillRankCounts.get(s.id) ?? 0;
-      return getSkillMaxDpPurchases(costElements, devType, tpRanks) > 0;
-    });
-  }, [skills, levellingSkillPurchases, categoryCostMap, skillDevTypeMap, tpGrantedSkillRankCounts]);
-
-  const levellingSkillCategoryOptions = useMemo((): RichSelectOption[] => {
-    const catIds = new Set(levellingAvailableSkills.map((s) => s.category));
-    return Array.from(catIds)
-      .map((id): RichSelectOption => {
-        const costElements = categoryCostMap.get(id) ?? [];
-        const firstRankCost = getSkillDpCostWithTpOffset(costElements, undefined, 1, 0);
-        const name = categoryNameById.get(id) ?? id;
-        const secondary = `${firstRankCost} DP`;
-        return {
-          value: id,
-          label: <RichOptionLabel primary={name} secondary={secondary} />,
-          searchText: `${name} \u2014 ${secondary}`,
-        };
-      })
-      .sort((a, b) => (a.searchText ?? '').localeCompare(b.searchText ?? ''));
-  }, [levellingAvailableSkills, categoryNameById, categoryCostMap]);
-
-  const levellingSkillOptions = useMemo(() => {
-    return levellingAvailableSkills
-      .filter((s) => !levellingSkillCategoryFilter || s.category === levellingSkillCategoryFilter)
-      .sort((a, b) => a.name.localeCompare(b.name))
-      .map((s): RichSelectOption => {
-        const costElements = categoryCostMap.get(s.category) ?? [];
-        const devType = skillDevTypeMap.get(s.id);
-        const tpRanks = tpGrantedSkillRankCounts.get(s.id) ?? 0;
-        const nextRankCost = getSkillDpCostWithTpOffset(costElements, devType, 1, tpRanks);
-        const secondary = devType ? `${devType} \u2014 ${nextRankCost} DP` : `${nextRankCost} DP`;
-        return {
-          value: s.id,
-          label: <RichOptionLabel primary={s.name} secondary={secondary} />,
-          searchText: `${s.name} ${secondary}`,
-        };
-      });
-  }, [levellingAvailableSkills, levellingSkillCategoryFilter, categoryCostMap, skillDevTypeMap, tpGrantedSkillRankCounts]);
-
-  const levellingCategoryOptions = useMemo(() => {
-    const selectedSet = new Set(levellingCategoryPurchases.map((p) => p.id));
-    return categories
-      .filter((c) => {
-        if (selectedSet.has(c.id)) return false;
-        const costElements = categoryCostMap.get(c.id) ?? [];
-        const tpRanks = tpGrantedCategoryRankCounts.get(c.id) ?? 0;
-        return getCategoryMaxDpPurchases(costElements, tpRanks) > 0;
-      })
-      .sort((a, b) => {
-        const aLabel = categoryNameById.get(a.id) ?? a.id;
-        const bLabel = categoryNameById.get(b.id) ?? b.id;
-        return aLabel.localeCompare(bLabel);
-      })
-      .map((c): RichSelectOption => {
-        const costElements = categoryCostMap.get(c.id) ?? [];
-        const tpRanks = tpGrantedCategoryRankCounts.get(c.id) ?? 0;
-        const nextRankCost = getCategoryDpCostWithTpOffset(costElements, 1, tpRanks);
-        const baseLabel = categoryNameById.get(c.id) ?? c.id;
-        const secondary = `${nextRankCost} DP`;
-        return {
-          value: c.id,
-          label: <RichOptionLabel primary={baseLabel} secondary={secondary} />,
-          searchText: `${baseLabel} \u2014 ${secondary}`,
-        };
-      });
-  }, [categories, levellingCategoryPurchases, categoryCostMap, categoryNameById, tpGrantedCategoryRankCounts]);
-
-  const levellingSpellCategoryOptions = useMemo(
-    () => characterBuilder.categorySpellLists
-      .map((c): RichSelectOption => {
-        const costElements = categoryCostMap.get(c.category) ?? [];
-        const nextRankCost = getCategoryOrSpellListPurchaseTotalCost(costElements, 1);
-        const catName = categoryNameById.get(c.category) ?? c.category;
-        const secondary = `${nextRankCost} DP / rank`;
-        return {
-          value: c.category,
-          label: <RichOptionLabel primary={catName} secondary={secondary} />,
-          searchText: `${catName} \u2014 ${secondary}`,
-        };
-      })
-      .sort((a, b) => (a.searchText ?? '').localeCompare(b.searchText ?? '')),
-    [characterBuilder.categorySpellLists, categoryNameById, categoryCostMap],
-  );
-
-  const levellingSpellListsInSelectedCategory = useMemo(() => {
-    if (!levellingSelectedSpellCategory) return [];
-    const catEntry = characterBuilder.categorySpellLists.find((c) => c.category === levellingSelectedSpellCategory);
-    if (!catEntry) return [];
-    const purchasedIds = new Set(levellingSpellListPurchases.map((p) => p.id));
-    return catEntry.spellLists
-      .filter((slId) => !purchasedIds.has(slId))
-      .map((slId) => ({ id: slId, name: spellListNameById.get(slId) ?? slId }))
-      .sort((a, b) => a.name.localeCompare(b.name));
-  }, [levellingSelectedSpellCategory, characterBuilder.categorySpellLists, spellListNameById, levellingSpellListPurchases]);
 
   useEffect(() => {
     if (!cultureTypeId) {
@@ -1778,12 +1175,11 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
           stat: roll.assignedStat,
           temporary: Number(roll.temporary) || 0,
           potential: roll.potential ?? 0,
-          racialBonus: (race?.statBonuses.find((b) => b.id === roll.assignedStat)?.value ?? 0)
-            + levellingStatGains.filter((s) => s === roll.assignedStat).length,
+          racialBonus: (race?.statBonuses.find((b) => b.id === roll.assignedStat)?.value ?? 0),
           totalBonus: 0,
         })),
     }));
-  }, [characterName, isMale, isPC, raceId, cultureTypeId, cultureId, professionId, selectedRealms, statRolls, race, levellingStatGains]);
+  }, [characterName, isMale, isPC, raceId, cultureTypeId, cultureId, professionId, selectedRealms, statRolls, race]);
 
   useEffect(() => {
     setCharacterBuilder((prev) => ({
@@ -1955,109 +1351,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
       languageAbilities: mappedBackgroundLanguages,
     }));
   }, [backgroundState]);
-
-  useEffect(() => {
-    setTpResolutions((prev) => {
-      const resMap = new Map(prev.map((r) => [r.tpId, r]));
-      return selectedLevellingTrainingPackages.map((tp) => resMap.get(tp.id) ?? createEmptyTpResolution(tp));
-    });
-  }, [selectedLevellingTrainingPackages]);
-
-  useEffect(() => {
-    // Aggregate ranks from all selected training packages including resolved TP choices
-    const tpSkillRanks: Array<{ id: string; subcategory?: string; value: number }> = [];
-    const tpCategoryRanks: Array<{ id: string; value: number }> = [];
-    const tpSpellListRanks: Array<{ id: string; value: number }> = [];
-
-    for (const tp of selectedLevellingTrainingPackages) {
-      for (const rank of tp.skillRanks ?? []) {
-        tpSkillRanks.push({ id: rank.id, ...(rank.subcategory ? { subcategory: rank.subcategory } : {}), value: rank.value });
-      }
-      for (const rank of tp.categoryRanks ?? []) {
-        tpCategoryRanks.push({ id: rank.id, value: rank.value });
-      }
-      for (const rank of (tp.spellListRanks ?? []).filter((r) => r.numChoices <= 0)) {
-        if (rank.optionalCategory) {
-          tpSpellListRanks.push({ id: rank.optionalCategory, value: rank.value });
-        }
-      }
-      // Include resolved choices
-      const res = tpResolutions.find((r) => r.tpId === tp.id);
-      if (res) {
-        for (let gi = 0; gi < (tp.skillRankChoices ?? []).length; gi++) {
-          for (const alloc of res.skillRankChoices[gi] ?? []) {
-            if (alloc.id) tpSkillRanks.push({ id: alloc.id, ...(alloc.subcategory ? { subcategory: alloc.subcategory } : {}), value: alloc.ranks });
-          }
-        }
-        for (let gi = 0; gi < (tp.categoryMultiSkillRankChoices ?? []).length; gi++) {
-          for (const alloc of res.categoryMultiSkillChoices[gi] ?? []) {
-            if (alloc.id) tpSkillRanks.push({ id: alloc.id, ...(alloc.subcategory ? { subcategory: alloc.subcategory } : {}), value: alloc.ranks });
-          }
-        }
-        for (let gi = 0; gi < (tp.groupMultiSkillRankChoices ?? []).length; gi++) {
-          for (const alloc of res.groupMultiSkillChoices[gi] ?? []) {
-            if (alloc.id) tpSkillRanks.push({ id: alloc.id, ...(alloc.subcategory ? { subcategory: alloc.subcategory } : {}), value: alloc.ranks });
-          }
-        }
-        for (let gi = 0; gi < (tp.groupCategoryAndSkillRankChoices ?? []).length; gi++) {
-          const choiceDef = (tp.groupCategoryAndSkillRankChoices ?? [])[gi];
-          if (!choiceDef) continue;
-          const slot = res.groupCategoryAndSkillChoices[gi];
-          if (!slot) continue;
-          if (slot.categoryId) tpCategoryRanks.push({ id: slot.categoryId, value: choiceDef.value });
-          if (slot.skillId) tpSkillRanks.push({ id: slot.skillId, ...(slot.subcategory ? { subcategory: slot.subcategory } : {}), value: choiceDef.value });
-        }
-        const choiceableSlRanks = (tp.spellListRanks ?? []).filter((r) => r.numChoices > 0);
-        for (let gi = 0; gi < choiceableSlRanks.length; gi++) {
-          for (const alloc of res.spellListChoices[gi] ?? []) {
-            if (alloc.id) tpSpellListRanks.push({ id: alloc.id, value: alloc.ranks });
-          }
-        }
-        for (let gi = 0; gi < (tp.spellListCategoryRankChoices ?? []).length; gi++) {
-          for (const alloc of res.spellListCategoryChoices[gi] ?? []) {
-            if (alloc.id) tpSpellListRanks.push({ id: alloc.id, value: alloc.ranks });
-          }
-        }
-      }
-    }
-
-    // Add levelling DP-purchased skill ranks
-    for (const p of levellingSkillPurchases) {
-      const categoryId = skillCategoryMap.get(p.id);
-      if (!categoryId) continue;
-      const devType = skillDevTypeMap.get(p.id);
-      const ranksPerPurchase = getSkillRanksPerPurchase(devType);
-      const totalRanks = devType === 'Restricted'
-        ? p.purchases  // each purchase = 1 rank for Restricted
-        : p.purchases * ranksPerPurchase;
-      if (totalRanks > 0) {
-        tpSkillRanks.push({ id: p.id, ...(p.subcategory ? { subcategory: p.subcategory } : {}), value: totalRanks });
-      }
-    }
-
-    // Add levelling DP-purchased category ranks
-    for (const p of levellingCategoryPurchases) {
-      if (p.purchases > 0) {
-        tpCategoryRanks.push({ id: p.id, value: p.purchases });
-      }
-    }
-
-    // Add levelling DP-purchased spell list ranks
-    for (const p of levellingSpellListPurchases) {
-      if (p.purchases > 0) {
-        tpSpellListRanks.push({ id: p.id, value: p.purchases });
-      }
-    }
-
-    setCharacterBuilder((prev) => ({
-      ...prev,
-      skillRanks: tpSkillRanks,
-      categoryRanks: tpCategoryRanks,
-      spellListRanks: tpSpellListRanks,
-      numHobbySkillRanks: (prev.hobbySkillRanks ?? []).reduce((sum, row) => sum + row.value, 0),
-      numAdolescentSpellListRanks: tpSpellListRanks.reduce((sum, row) => sum + row.value, 0),
-    }));
-  }, [selectedLevellingTrainingPackages, levellingSkillPurchases, levellingCategoryPurchases, levellingSpellListPurchases, skillCategoryMap, skillDevTypeMap, tpResolutions]);
 
   const validateInitial = (): string | undefined => {
     if (!characterName.trim()) return 'Name is required.';
@@ -2456,55 +1749,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
     return undefined;
   };
 
-  const validateApprenticeship = (): string | undefined => {
-    if (levellingTotalDpSpent > characterBuilder.developmentPoints) {
-      return `Development points overspent by ${levellingTotalDpSpent - characterBuilder.developmentPoints}.`;
-    }
-
-    // Require all TP choices to be resolved before completing apprenticeship
-    if (tpsRequiringResolution.length > 0 && levellingSubstep !== 'purchasing') {
-      return 'Resolve all training package choices before completing apprenticeship.';
-    }
-    for (const tp of tpsRequiringResolution) {
-      const res = tpResolutions.find((r) => r.tpId === tp.id);
-      if (!res) return `Training package choices not complete for ${tp.name}.`;
-      const resError = validateTpResolution(tp, res);
-      if (resError) return resError;
-    }
-
-    // Check only one lifestyle package selected
-    let lifestyleCount = 0;
-    for (const tpId of levellingTrainingPackageIds) {
-      const tp = trainingPackages.find((t) => t.id === tpId);
-      if (tp?.lifestyle) lifestyleCount++;
-    }
-    if (lifestyleCount > 1) return 'Only one lifestyle training package may be selected.';
-
-    // Validate skill purchase subcategories
-    for (const p of levellingSkillPurchases) {
-      if (p.purchases <= 0) continue;
-      const skillName = skillNameById.get(p.id) ?? p.id;
-      const isWeaponGroupSkill = weaponGroupSkillIds.has(p.id);
-
-      if (isWeaponGroupSkill && p.subcategory) {
-        const validWeaponTypes = weaponTypeOptionsBySkillId.get(p.id) ?? [];
-        if (!validWeaponTypes.some((wt) => wt.value === p.subcategory)) {
-          return `Levelling Skill: invalid weapon type selected for ${skillName}.`;
-        }
-      }
-
-      if (!isWeaponGroupSkill && mandatorySubcategorySkillIds.has(p.id) && !p.subcategory.trim()) {
-        return `Levelling Skill: enter subcategory for ${skillName}.`;
-      }
-
-      if (isWeaponGroupSkill && mandatorySubcategorySkillIds.has(p.id) && !p.subcategory) {
-        return `Levelling Skill: select weapon type for ${skillName}.`;
-      }
-    }
-
-    return undefined;
-  };
-
   const validatePhysique = (): string | undefined => {
     if (!physiqueAutoHeight) {
       if (!physiqueEnteredHeightStr.trim()) return 'Enter height in inches or enable auto-generate.';
@@ -2526,7 +1770,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
       physique: validatePhysique(),
       hobby: validateHobby(),
       background: validateBackground(),
-      apprenticeship: validateApprenticeship(),
     };
     setErrors(next);
     return next;
@@ -2582,14 +1825,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
     backgroundLanguageRankSpent,
     backgroundLanguageRanksBudget,
     backgroundLanguageRankRemaining,
-    levellingTrainingPackageIds,
-    levellingStatGains,
-    levellingSkillPurchases,
-    levellingCategoryPurchases,
-    levellingSpellListPurchases,
-    levellingTotalDpSpent,
-    levellingSubstep,
-    tpResolutions,
   ]);
 
   const canGoNext = (() => {
@@ -2725,8 +1960,7 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
             stat,
             temporary: Number(assigned.temporary) || 0,
             potential: assigned.potential ?? 0,
-            racialBonus: (race?.statBonuses.find((b) => b.id === stat)?.value ?? 0)
-              + levellingStatGains.filter((s) => s === stat).length,
+            racialBonus: (race?.statBonuses.find((b) => b.id === stat)?.value ?? 0),
             totalBonus: 0,
           };
         });
@@ -2968,11 +2202,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
         );
 
         setCharacterBuilder(response);
-        setPreLevellingRanks({
-          skillRanks: response.skillRanks ?? [],
-          categoryRanks: response.categoryRanks ?? [],
-          spellListRanks: response.spellListRanks ?? [],
-        });
       } catch (e) {
         toast({
           variant: 'danger',
@@ -3268,209 +2497,8 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
     setHobbySpellListOptions([]);
     setHobbySpellListId('');
     resetBackgroundState();
-    setLevellingTrainingPackageIds([]);
-    setLevellingStatGains([]);
-    setLevellingSkillPurchases([]);
-    setLevellingSkillCategoryFilter('');
-    setLevellingSkillPendingId('');
-    setLevellingSkillPendingSubcategory('');
-    setLevellingCategoryPurchases([]);
-    setLevellingSpellListPurchases([]);
-    setLevellingSelectedSpellCategory('');
-    setLevellingAddingSpellList(false);
-    setLevellingSubstep('selecting');
-    setLevellingResolvingTpIndex(0);
-    setTpResolutions([]);
-    setPreLevellingRanks({ skillRanks: [], categoryRanks: [], spellListRanks: [] });
     setCharacterBuilder(createEmptyCharacterBuilder());
     setErrors({});
-  };
-
-  const submitLevelUpgrade = async (): Promise<boolean> => {
-    if (!raceId || !cultureId || !professionId || selectedRealms.length === 0) return false;
-
-    const nextErrors = recomputeErrors();
-    if (Object.values(nextErrors).some(Boolean)) {
-      toast({
-        variant: 'warning',
-        title: 'Cannot apply level upgrade',
-        description: 'Please resolve validation issues in earlier steps.',
-      });
-      return false;
-    }
-
-    setApplying(true);
-    try {
-      const totalStatGains: Stat[] = [];
-      for (const tp of selectedLevellingTrainingPackages) {
-        for (const stat of tp.statGains ?? []) totalStatGains.push(stat);
-        const res = tpResolutions.find((r) => r.tpId === tp.id);
-        if (res) {
-          for (const stat of res.statGainChoices) {
-            if (stat) totalStatGains.push(stat);
-          }
-        }
-      }
-      totalStatGains.push(...levellingStatGains);
-
-      const totalSkillsByKey = new Map<string, { id: string; subcategory?: string | undefined; ranks: number }>();
-      const addSkillRanks = (id: string, subcategory: string | undefined, ranks: number) => {
-        if (!id || ranks <= 0) return;
-        const normalizedSubcategory = subcategory?.trim() || undefined;
-        const key = skillChoiceKey(id, normalizedSubcategory);
-        const existing = totalSkillsByKey.get(key);
-        if (existing) {
-          existing.ranks += ranks;
-        } else {
-          totalSkillsByKey.set(key, {
-            id,
-            ...(normalizedSubcategory ? { subcategory: normalizedSubcategory } : {}),
-            ranks,
-          });
-        }
-      };
-      for (const rank of preLevellingRanks.skillRanks ?? []) {
-        addSkillRanks(rank.id, rank.subcategory, rank.value ?? 0);
-      }
-      for (const rank of characterBuilder.skillRanks ?? []) {
-        addSkillRanks(rank.id, rank.subcategory, rank.value ?? 0);
-      }
-
-      const totalCategoryRanks = new Map<string, number>();
-      const addCategoryRanks = (id: string, ranks: number) => {
-        if (!id || ranks <= 0) return;
-        totalCategoryRanks.set(id, (totalCategoryRanks.get(id) ?? 0) + ranks);
-      };
-      for (const rank of preLevellingRanks.categoryRanks ?? []) {
-        addCategoryRanks(rank.id, rank.value ?? 0);
-      }
-      for (const rank of characterBuilder.categoryRanks ?? []) {
-        addCategoryRanks(rank.id, rank.value ?? 0);
-      }
-
-      const knownSpellListIds = new Set(spellLists.map((sl) => sl.id));
-      const spellListTotals = new Map<string, number>();
-      const spellListCategoryOverrides = new Map<string, string>();
-      const addSpellListRanks = (spellListId: string, ranks: number) => {
-        if (!spellListId || ranks <= 0 || !knownSpellListIds.has(spellListId)) return;
-        spellListTotals.set(spellListId, (spellListTotals.get(spellListId) ?? 0) + ranks);
-      };
-
-      for (const rank of preLevellingRanks.spellListRanks ?? []) {
-        addSpellListRanks(rank.id, rank.value ?? 0);
-      }
-
-      for (const p of levellingSpellListPurchases) {
-        if (p.purchases > 0) addSpellListRanks(p.id, p.purchases);
-      }
-
-      for (const tp of selectedLevellingTrainingPackages) {
-        const res = tpResolutions.find((r) => r.tpId === tp.id);
-
-        const fixedSpellRows = (tp.spellListRanks ?? []).filter((r) => r.numChoices <= 0);
-        for (const row of fixedSpellRows) {
-          for (const spellListId of row.options ?? []) {
-            addSpellListRanks(spellListId, row.value);
-            if (row.optionalCategory?.trim()) {
-              spellListCategoryOverrides.set(spellListId, row.optionalCategory.trim());
-            }
-          }
-        }
-
-        const choiceableSpellRows = (tp.spellListRanks ?? []).filter((r) => r.numChoices > 0);
-        for (let gi = 0; gi < choiceableSpellRows.length; gi++) {
-          const row = choiceableSpellRows[gi];
-          if (!row) continue;
-          const allocs = res?.spellListChoices[gi] ?? [];
-          for (const alloc of allocs) {
-            addSpellListRanks(alloc.id, alloc.ranks);
-            if (alloc.id && row.optionalCategory?.trim()) {
-              spellListCategoryOverrides.set(alloc.id, row.optionalCategory.trim());
-            }
-          }
-        }
-
-        for (let gi = 0; gi < (tp.spellListCategoryRankChoices ?? []).length; gi++) {
-          const allocs = res?.spellListCategoryChoices[gi] ?? [];
-          for (const alloc of allocs) {
-            addSpellListRanks(alloc.id, alloc.ranks);
-          }
-        }
-      }
-
-      const totalLanguages = new Map<string, { language: string; spoken: number; written: number; somatic: number }>();
-      const ensureLanguageRow = (languageId: string) => {
-        let row = totalLanguages.get(languageId);
-        if (!row) {
-          row = { language: languageId, spoken: 0, written: 0, somatic: 0 };
-          totalLanguages.set(languageId, row);
-        }
-        return row;
-      };
-      for (const row of characterBuilder.languageAbilities ?? []) {
-        if (!row.language) continue;
-        const target = ensureLanguageRow(row.language);
-        target.spoken += Math.max(0, row.spoken ?? 0);
-        target.written += Math.max(0, row.written ?? 0);
-        target.somatic += Math.max(0, row.somatic ?? 0);
-      }
-      for (const res of tpResolutions) {
-        for (const group of res.languageChoices ?? []) {
-          for (const alloc of group ?? []) {
-            if (!alloc.languageId) continue;
-            const target = ensureLanguageRow(alloc.languageId);
-            target.spoken += Math.max(0, alloc.spoken ?? 0);
-            target.written += Math.max(0, alloc.written ?? 0);
-            target.somatic += Math.max(0, alloc.somatic ?? 0);
-          }
-        }
-      }
-
-      const payload: CharacterBuilder = {
-        ...characterBuilder,
-        levellingTrainingPackages: levellingTrainingPackageIds,
-        levellingStatGains: totalStatGains,
-        skillRanks: Array.from(totalSkillsByKey.values())
-          .filter((row) => row.ranks > 0)
-          .sort((a, b) => skillChoiceKey(a.id, a.subcategory).localeCompare(skillChoiceKey(b.id, b.subcategory)))
-          .map(({ id, subcategory, ranks }) => ({ id, ...(subcategory ? { subcategory } : {}), value: ranks })),
-        categoryRanks: Array.from(totalCategoryRanks.entries())
-          .filter(([, ranks]) => ranks > 0)
-          .map(([id, ranks]) => ({ id, value: ranks }))
-          .sort((a, b) => a.id.localeCompare(b.id)),
-        spellListRanks: Array.from(spellListTotals.entries())
-          .filter(([, ranks]) => ranks > 0)
-          .map(([id, ranks]) => ({ id, value: ranks }))
-          .sort((a, b) => a.id.localeCompare(b.id)),
-        languageAbilities: Array.from(totalLanguages.values())
-          .filter((row) => row.spoken > 0 || row.written > 0 || row.somatic > 0)
-          .map((row) => ({
-            language: row.language,
-            ...(row.spoken > 0 ? { spoken: row.spoken } : {}),
-            ...(row.written > 0 ? { written: row.written } : {}),
-            ...(row.somatic > 0 ? { somatic: row.somatic } : {}),
-          }))
-          .sort((a, b) => a.language.localeCompare(b.language)),
-      };
-
-      const response = await applyLevellingChoices(payload);
-      setCharacterBuilder(response);
-      toast({
-        variant: 'success',
-        title: 'Apprenticeship choices applied',
-        description: 'Your character has been successfully built.',
-      });
-      return true;
-    } catch (e) {
-      toast({
-        variant: 'danger',
-        title: 'Apply apprenticeship choices failed',
-        description: String(e instanceof Error ? e.message : e),
-      });
-      return false;
-    } finally {
-      setApplying(false);
-    }
   };
 
   if (loading) return <div>Loading character workflow…</div>;
@@ -3612,7 +2640,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
                   value={raceId}
                   onChange={(v) => {
                     setRaceId(v);
-                    setLevellingTrainingPackageIds([]);
                     resetBackgroundState();
                   }}
                   options={races
@@ -3627,7 +2654,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
                   value={cultureTypeId}
                   onChange={(v) => {
                     setCultureTypeId(v);
-                    setLevellingTrainingPackageIds([]);
                     resetBackgroundState();
                   }}
                   options={cultureTypes
@@ -3642,7 +2668,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
                   value={cultureId}
                   onChange={(v) => {
                     setCultureId(v);
-                    setLevellingTrainingPackageIds([]);
                   }}
                   options={availableCultures
                     .slice()
@@ -4783,1053 +3808,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
             </section>
           )}
 
-          {step === 'apprenticeship' && (
-            <section style={{ display: 'grid', gap: 12 }}>
-              <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-                <div style={{ color: 'var(--muted)' }}>
-                  {levellingSubstep === 'selecting' && 'Select training packages. TPs with required choices must be resolved before spending extra development points.'}
-                  {levellingSubstep === 'resolving' && `Resolve training package choices (${levellingResolvingTpIndex + 1} of ${tpsRequiringResolution.length}).`}
-                  {levellingSubstep === 'purchasing' && 'Spend remaining development points on stat gains, skill ranks, category ranks, and spell list ranks. Unspent points carry to the next level.'}
-                </div>
-                <div style={{ fontWeight: 600, whiteSpace: 'nowrap', marginLeft: 16 }}>
-                  DP: {levellingDpRemaining} / {characterBuilder.developmentPoints}
-                </div>
-              </div>
-
-              {/* SELECTING sub-step */}
-              {levellingSubstep === 'selecting' && (
-                <>
-                  <div style={{ border: '1px solid var(--border)', borderRadius: 8, padding: 10 }}>
-                    <h4 style={{ margin: '0 0 8px' }}>Training Packages</h4>
-                    {selectedLevellingTrainingPackages.length > 0 && (
-                      <div style={{ display: 'grid', gap: 6, marginBottom: 8 }}>
-                        {selectedLevellingTrainingPackages.map((tp) => {
-                          const cost = tpCostMap.get(tp.id) ?? 0;
-                          return (
-                            <div key={tp.id} style={{ display: 'flex', gap: 8, alignItems: 'center' }}>
-                              <span style={{ flex: 1 }}>
-                                {tp.name}{tp.lifestyle ? ' (Lifestyle)' : ''} — {cost} DP
-                                {tpHasChoices(tp) && <span style={{ color: 'var(--muted)', marginLeft: 6 }}>(choices required)</span>}
-                              </span>
-                              <button
-                                type="button"
-                                onClick={() => setLevellingTrainingPackageIds((prev) => prev.filter((id) => id !== tp.id))}
-                              >
-                                Remove
-                              </button>
-                            </div>
-                          );
-                        })}
-                      </div>
-                    )}
-                    <RichSelect
-                      label="Add Training Package"
-                      hideLabel={true}
-                      value=""
-                      onChange={(v) => {
-                        if (v) setLevellingTrainingPackageIds((prev) => [...prev, v]);
-                      }}
-                      options={levellingTrainingPackageOptions}
-                      placeholderOption="— Add training package —"
-                    />
-                  </div>
-                  <div>
-                    <button
-                      type="button"
-                      onClick={() => {
-                        if (tpsRequiringResolution.length === 0) {
-                          setLevellingSubstep('purchasing');
-                        } else {
-                          setLevellingResolvingTpIndex(0);
-                          setLevellingSubstep('resolving');
-                        }
-                      }}
-                    >
-                      {tpsRequiringResolution.length === 0 ? 'Proceed to Extra Purchases ?' : 'Proceed to Resolve TP Choices ?'}
-                    </button>
-                  </div>
-                </>
-              )}
-
-              {/* RESOLVING sub-step */}
-              {levellingSubstep === 'resolving' && (() => {
-                const currentTp = tpsRequiringResolution[levellingResolvingTpIndex];
-                const currentResolution = currentTp ? tpResolutions.find((r) => r.tpId === currentTp.id) : undefined;
-                if (!currentTp || !currentResolution) return null;
-                const tpRealmStatGains = new Set<Stat>(
-                  currentTp.realmStatGain
-                    ? characterBuilder.magicalRealms.flatMap((realm) => {
-                      const stat = getStatForRealm(realm);
-                      return stat ? [stat] : [];
-                    })
-                    : [],
-                );
-                const resError = validateTpResolution(currentTp, currentResolution, tpRealmStatGains);
-                const updateResolution = (updater: (r: TpResolution) => TpResolution) => {
-                  setTpResolutions((prev) => prev.map((r) => r.tpId === currentTp.id ? updater(r) : r));
-                };
-                return (
-                  <div style={{ display: 'grid', gap: 10 }}>
-                    <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-                      <h4 style={{ margin: 0 }}>{currentTp.name} — {tpCostMap.get(currentTp.id) ?? 0} DP</h4>
-                      <span style={{ color: 'var(--muted)' }}>TP {levellingResolvingTpIndex + 1} of {tpsRequiringResolution.length}</span>
-                    </div>
-
-                    {currentTp.realmStatGain && characterBuilder.magicalRealms.length > 0 && (
-                      <div style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                        <strong>Realm Stat Gain</strong>
-                        <div style={{ color: 'var(--muted)', fontSize: '0.9em', marginTop: 4 }}>
-                          This TP grants a stat gain roll for the following stats: {[...tpRealmStatGains].join(', ')}.
-                        </div>
-                      </div>
-                    )}
-
-                    {currentTp.statGainChoices && currentTp.statGainChoices.numChoices > 0 && (
-                      <div style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                        <strong>Stat Gain Choices</strong>
-                        <div style={{ color: 'var(--muted)', fontSize: '0.9em', marginBottom: 8 }}>
-                          Choose {currentTp.statGainChoices.numChoices} stat{currentTp.statGainChoices.numChoices > 1 ? 's' : ''} to receive a gain roll.
-                        </div>
-                        <div style={{ display: 'grid', gap: 6 }}>
-                          {Array.from({ length: currentTp.statGainChoices.numChoices }, (_, si) => {
-                            const chosen = currentResolution.statGainChoices[si] ?? '';
-                            const otherChosen = new Set<string>(currentResolution.statGainChoices.filter((s, i) => i !== si && s));
-                            const opts = (currentTp.statGainChoices!.options as string[])
-                              .filter((s) => !otherChosen.has(s) && !tpRealmStatGains.has(s as Stat))
-                              .map((s) => ({ value: s, label: s }));
-                            return (
-                              <LabeledSelect
-                                key={si}
-                                label={`Stat ${si + 1}`}
-                                value={chosen}
-                                onChange={(v) => updateResolution((r) => ({
-                                  ...r,
-                                  statGainChoices: r.statGainChoices.map((s, i) => i === si ? v as Stat | '' : s),
-                                }))}
-                                options={opts}
-                                placeholderOption="— Select stat —"
-                              />
-                            );
-                          })}
-                        </div>
-                      </div>
-                    )}
-
-                    {(currentTp.skillRankChoices ?? []).some((c) => c.numChoices > 0) && (
-                      <div style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                        <strong>Skill Rank Choices</strong>
-                        <div style={{ display: 'grid', gap: 10, marginTop: 6 }}>
-                          {(currentTp.skillRankChoices ?? []).map((choiceDef, gi) => {
-                            if (choiceDef.numChoices <= 0) return null;
-                            const allocs = currentResolution.skillRankChoices[gi] ?? [];
-                            const totalAllocated = allocs.reduce((s, a) => s + a.ranks, 0);
-                            const remainingRanks = choiceDef.value - totalAllocated;
-                            const usedIds = new Set(allocs.map((a) => a.id).filter(Boolean));
-                            const optionIds = new Set(choiceDef.options.map((o) => o.id));
-                            return (
-                              <div key={gi}>
-                                <div style={{ color: 'var(--muted)', fontSize: '0.9em', marginBottom: 4 }}>
-                                  Distribute {choiceDef.value} rank{choiceDef.value !== 1 ? 's' : ''} across up to {choiceDef.numChoices} skill{choiceDef.numChoices !== 1 ? 's' : ''}
-                                  {remainingRanks !== 0 && <span style={{ color: remainingRanks > 0 ? 'var(--warning, #b96c00)' : '#b00020', marginLeft: 8, fontWeight: 600 }}>({remainingRanks > 0 ? `${remainingRanks} rank${remainingRanks !== 1 ? 's' : ''} unallocated` : `${Math.abs(remainingRanks)} rank${Math.abs(remainingRanks) !== 1 ? 's' : ''} over`})</span>}
-                                </div>
-                                <div style={{ display: 'grid', gap: 6 }}>
-                                  {allocs.map((alloc, ai) => {
-                                    const availableOpts = skills.filter((s) => optionIds.has(s.id) && (s.id === alloc.id || !usedIds.has(s.id))).map(toSkillRichOption);
-                                    return (
-                                      <div key={ai} style={{ display: 'flex', gap: 6, alignItems: 'flex-end' }}>
-                                        <div style={{ flex: 1 }}>
-                                          <RichSelect
-                                            label={`Skill ${ai + 1}`}
-                                            value={alloc.id}
-                                            onChange={(v) => updateResolution((r) => ({ ...r, skillRankChoices: r.skillRankChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, id: v, subcategory: '' })) }))}
-                                            options={availableOpts}
-                                            placeholderOption="— Select skill —"
-                                          />
-                                        </div>
-                                        <div style={{ display: 'flex', alignItems: 'center', gap: 4, paddingBottom: 4 }}>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, skillRankChoices: r.skillRankChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, ranks: a.ranks - 1 })) }))} disabled={alloc.ranks <= 1}>-</button>
-                                          <span style={{ minWidth: 24, textAlign: 'center' }}>{alloc.ranks}</span>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, skillRankChoices: r.skillRankChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, ranks: a.ranks + 1 })) }))} disabled={remainingRanks <= 0}>+</button>
-                                        </div>
-                                        <button type="button" onClick={() => updateResolution((r) => ({ ...r, skillRankChoices: r.skillRankChoices.map((g, gIdx) => gIdx !== gi ? g : g.filter((_, aIdx) => aIdx !== ai)) }))} style={{ marginBottom: 4 }}>×</button>
-                                      </div>
-                                    );
-                                  })}
-                                </div>
-                                {allocs.length < choiceDef.numChoices && remainingRanks > 0 && (
-                                  <button type="button" style={{ marginTop: 4 }} onClick={() => updateResolution((r) => ({ ...r, skillRankChoices: r.skillRankChoices.map((g, gIdx) => gIdx !== gi ? g : [...g, { id: '', subcategory: '', ranks: 1 }]) }))}>+ Add skill</button>
-                                )}
-                              </div>
-                            );
-                          })}
-                        </div>
-                      </div>
-                    )}
-
-                    {(currentTp.categoryMultiSkillRankChoices ?? []).some((c) => c.numChoices > 0) && (
-                      <div style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                        <strong>Category Skill Choices</strong>
-                        <div style={{ display: 'grid', gap: 10, marginTop: 6 }}>
-                          {(currentTp.categoryMultiSkillRankChoices ?? []).map((choiceDef, gi) => {
-                            if (choiceDef.numChoices <= 0) return null;
-                            const catName = categoryNameById.get(choiceDef.id) ?? choiceDef.id;
-                            const catSkills = skillIdsByCategory.get(choiceDef.id) ?? [];
-                            const allocs = currentResolution.categoryMultiSkillChoices[gi] ?? [];
-                            const totalAllocated = allocs.reduce((s, a) => s + a.ranks, 0);
-                            const remainingRanks = choiceDef.value - totalAllocated;
-                            const usedIds = new Set(allocs.map((a) => a.id).filter(Boolean));
-                            return (
-                              <div key={gi}>
-                                <div style={{ color: 'var(--muted)', fontSize: '0.9em', marginBottom: 4 }}>
-                                  {catName}: distribute {choiceDef.value} rank{choiceDef.value !== 1 ? 's' : ''} across up to {choiceDef.numChoices} skill{choiceDef.numChoices !== 1 ? 's' : ''}
-                                  {remainingRanks !== 0 && <span style={{ color: remainingRanks > 0 ? 'var(--warning, #b96c00)' : '#b00020', marginLeft: 8, fontWeight: 600 }}>({remainingRanks > 0 ? `${remainingRanks} rank${remainingRanks !== 1 ? 's' : ''} unallocated` : `${Math.abs(remainingRanks)} rank${Math.abs(remainingRanks) !== 1 ? 's' : ''} over`})</span>}
-                                </div>
-                                <div style={{ display: 'grid', gap: 6 }}>
-                                  {allocs.map((alloc, ai) => {
-                                    const availableOpts = catSkills.filter((s) => s.id === alloc.id || !usedIds.has(s.id)).map(toSkillRichOption);
-                                    return (
-                                      <div key={ai} style={{ display: 'flex', gap: 6, alignItems: 'flex-end' }}>
-                                        <div style={{ flex: 1 }}>
-                                          <RichSelect
-                                            label={`Skill ${ai + 1}`}
-                                            value={alloc.id}
-                                            onChange={(v) => updateResolution((r) => ({ ...r, categoryMultiSkillChoices: r.categoryMultiSkillChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, id: v, subcategory: '' })) }))}
-                                            options={availableOpts}
-                                            placeholderOption="— Select skill —"
-                                          />
-                                        </div>
-                                        <div style={{ display: 'flex', alignItems: 'center', gap: 4, paddingBottom: 4 }}>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, categoryMultiSkillChoices: r.categoryMultiSkillChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, ranks: a.ranks - 1 })) }))} disabled={alloc.ranks <= 1}>-</button>
-                                          <span style={{ minWidth: 24, textAlign: 'center' }}>{alloc.ranks}</span>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, categoryMultiSkillChoices: r.categoryMultiSkillChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, ranks: a.ranks + 1 })) }))} disabled={remainingRanks <= 0}>+</button>
-                                        </div>
-                                        <button type="button" onClick={() => updateResolution((r) => ({ ...r, categoryMultiSkillChoices: r.categoryMultiSkillChoices.map((g, gIdx) => gIdx !== gi ? g : g.filter((_, aIdx) => aIdx !== ai)) }))} style={{ marginBottom: 4 }}>×</button>
-                                      </div>
-                                    );
-                                  })}
-                                </div>
-                                {allocs.length < choiceDef.numChoices && remainingRanks > 0 && (
-                                  <button type="button" style={{ marginTop: 4 }} onClick={() => updateResolution((r) => ({ ...r, categoryMultiSkillChoices: r.categoryMultiSkillChoices.map((g, gIdx) => gIdx !== gi ? g : [...g, { id: '', subcategory: '', ranks: 1 }]) }))}>+ Add skill</button>
-                                )}
-                              </div>
-                            );
-                          })}
-                        </div>
-                      </div>
-                    )}
-
-                    {(currentTp.groupMultiSkillRankChoices ?? []).some((c) => c.numChoices > 0) && (
-                      <div style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                        <strong>Group Skill Choices</strong>
-                        <div style={{ display: 'grid', gap: 10, marginTop: 6 }}>
-                          {(currentTp.groupMultiSkillRankChoices ?? []).map((choiceDef, gi) => {
-                            if (choiceDef.numChoices <= 0) return null;
-                            const grpName = groups.find((g) => g.id === choiceDef.id)?.name ?? choiceDef.id;
-                            const grpSkills = (categoryIdsByGroup.get(choiceDef.id) ?? []).flatMap((catId) => skillIdsByCategory.get(catId) ?? []);
-                            const allocs = currentResolution.groupMultiSkillChoices[gi] ?? [];
-                            const totalAllocated = allocs.reduce((s, a) => s + a.ranks, 0);
-                            const remainingRanks = choiceDef.value - totalAllocated;
-                            const usedIds = new Set(allocs.map((a) => a.id).filter(Boolean));
-                            return (
-                              <div key={gi}>
-                                <div style={{ color: 'var(--muted)', fontSize: '0.9em', marginBottom: 4 }}>
-                                  {grpName}: distribute {choiceDef.value} rank{choiceDef.value !== 1 ? 's' : ''} across up to {choiceDef.numChoices} skill{choiceDef.numChoices !== 1 ? 's' : ''}
-                                  {remainingRanks !== 0 && <span style={{ color: remainingRanks > 0 ? 'var(--warning, #b96c00)' : '#b00020', marginLeft: 8, fontWeight: 600 }}>({remainingRanks > 0 ? `${remainingRanks} rank${remainingRanks !== 1 ? 's' : ''} unallocated` : `${Math.abs(remainingRanks)} rank${Math.abs(remainingRanks) !== 1 ? 's' : ''} over`})</span>}
-                                </div>
-                                <div style={{ display: 'grid', gap: 6 }}>
-                                  {allocs.map((alloc, ai) => {
-                                    const availableOpts = grpSkills.filter((s) => s.id === alloc.id || !usedIds.has(s.id)).map(toSkillRichOption);
-                                    return (
-                                      <div key={ai} style={{ display: 'flex', gap: 6, alignItems: 'flex-end' }}>
-                                        <div style={{ flex: 1 }}>
-                                          <RichSelect
-                                            label={`Skill ${ai + 1}`}
-                                            value={alloc.id}
-                                            onChange={(v) => updateResolution((r) => ({ ...r, groupMultiSkillChoices: r.groupMultiSkillChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, id: v, subcategory: '' })) }))}
-                                            options={availableOpts}
-                                            placeholderOption="— Select skill —"
-                                          />
-                                        </div>
-                                        <div style={{ display: 'flex', alignItems: 'center', gap: 4, paddingBottom: 4 }}>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, groupMultiSkillChoices: r.groupMultiSkillChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, ranks: a.ranks - 1 })) }))} disabled={alloc.ranks <= 1}>-</button>
-                                          <span style={{ minWidth: 24, textAlign: 'center' }}>{alloc.ranks}</span>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, groupMultiSkillChoices: r.groupMultiSkillChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, ranks: a.ranks + 1 })) }))} disabled={remainingRanks <= 0}>+</button>
-                                        </div>
-                                        <button type="button" onClick={() => updateResolution((r) => ({ ...r, groupMultiSkillChoices: r.groupMultiSkillChoices.map((g, gIdx) => gIdx !== gi ? g : g.filter((_, aIdx) => aIdx !== ai)) }))} style={{ marginBottom: 4 }}>×</button>
-                                      </div>
-                                    );
-                                  })}
-                                </div>
-                                {allocs.length < choiceDef.numChoices && remainingRanks > 0 && (
-                                  <button type="button" style={{ marginTop: 4 }} onClick={() => updateResolution((r) => ({ ...r, groupMultiSkillChoices: r.groupMultiSkillChoices.map((g, gIdx) => gIdx !== gi ? g : [...g, { id: '', subcategory: '', ranks: 1 }]) }))}>+ Add skill</button>
-                                )}
-                              </div>
-                            );
-                          })}
-                        </div>
-                      </div>
-                    )}
-
-                    {(currentTp.groupCategoryAndSkillRankChoices ?? []).length > 0 && (
-                      <div style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                        <strong>Group Category &amp; Skill Choices</strong>
-                        <div style={{ display: 'grid', gap: 10, marginTop: 6 }}>
-                          {(currentTp.groupCategoryAndSkillRankChoices ?? []).map((choiceDef, gi) => {
-                            const slot = currentResolution.groupCategoryAndSkillChoices[gi] ?? { categoryId: '', skillId: '', subcategory: '' };
-                            const grpName = groups.find((g) => g.id === choiceDef.id)?.name ?? choiceDef.id;
-                            const isWeaponsGroup = choiceDef.id === 'SKILLGROUP_WEAPON';
-                            const catsInGroup = categories.filter((c) => c.group === choiceDef.id).map((c) => ({ value: c.id, label: c.name }));
-                            const skillsInSelectedCat = slot.categoryId ? (skillIdsByCategory.get(slot.categoryId) ?? []).map(toSkillRichOption) : [];
-                            const weaponTypeOpts = slot.skillId && isWeaponsGroup ? (weaponTypeOptionsBySkillId.get(slot.skillId) ?? []) : [];
-                            return (
-                              <div key={gi} style={{ display: 'grid', gap: 6 }}>
-                                <div style={{ color: 'var(--muted)', fontSize: '0.9em' }}>
-                                  {grpName}: select a category and one skill from it — each receives {choiceDef.value} rank{choiceDef.value !== 1 ? 's' : ''}
-                                </div>
-                                <LabeledSelect
-                                  label="Category"
-                                  value={slot.categoryId}
-                                  onChange={(v) => updateResolution((r) => ({
-                                    ...r,
-                                    groupCategoryAndSkillChoices: r.groupCategoryAndSkillChoices.map((s, i) => i !== gi ? s : { categoryId: v, skillId: '', subcategory: '' }),
-                                  }))}
-                                  options={catsInGroup}
-                                  placeholderOption="— Select category —"
-                                />
-                                {slot.categoryId && (
-                                  <RichSelect
-                                    label="Skill"
-                                    value={slot.skillId}
-                                    onChange={(v) => updateResolution((r) => ({
-                                      ...r,
-                                      groupCategoryAndSkillChoices: r.groupCategoryAndSkillChoices.map((s, i) => i !== gi ? s : { ...s, skillId: v, subcategory: '' }),
-                                    }))}
-                                    options={skillsInSelectedCat}
-                                    placeholderOption="— Select skill —"
-                                  />
-                                )}
-                                {slot.skillId && isWeaponsGroup && weaponTypeOpts.length > 0 && (
-                                  <LabeledSelect
-                                    label="Weapon Type"
-                                    value={slot.subcategory}
-                                    onChange={(v) => updateResolution((r) => ({
-                                      ...r,
-                                      groupCategoryAndSkillChoices: r.groupCategoryAndSkillChoices.map((s, i) => i !== gi ? s : { ...s, subcategory: v }),
-                                    }))}
-                                    options={weaponTypeOpts}
-                                    placeholderOption="— Select weapon type —"
-                                  />
-                                )}
-                              </div>
-                            );
-                          })}
-                        </div>
-                      </div>
-                    )}
-
-                    {(currentTp.spellListRanks ?? []).some((r) => r.numChoices > 0) && (
-                      <div style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                        <strong>Spell List Choices</strong>
-                        <div style={{ display: 'grid', gap: 10, marginTop: 6 }}>
-                          {(currentTp.spellListRanks ?? []).filter((r) => r.numChoices > 0).map((choiceDef, gi) => {
-                            const allocs = currentResolution.spellListChoices[gi] ?? [];
-                            const totalAllocated = allocs.reduce((s, a) => s + a.ranks, 0);
-                            const remainingRanks = choiceDef.value - totalAllocated;
-                            const usedIds = new Set(allocs.map((a) => a.id).filter(Boolean));
-                            const availableSlOpts = choiceDef.options.filter((slId) => slId).map(toSpellListRichOption);
-                            return (
-                              <div key={gi}>
-                                <div style={{ color: 'var(--muted)', fontSize: '0.9em', marginBottom: 4 }}>
-                                  Distribute {choiceDef.value} rank{choiceDef.value !== 1 ? 's' : ''} across up to {choiceDef.numChoices} spell list{choiceDef.numChoices !== 1 ? 's' : ''}
-                                  {remainingRanks !== 0 && <span style={{ color: remainingRanks > 0 ? 'var(--warning, #b96c00)' : '#b00020', marginLeft: 8, fontWeight: 600 }}>({remainingRanks > 0 ? `${remainingRanks} rank${remainingRanks !== 1 ? 's' : ''} unallocated` : `${Math.abs(remainingRanks)} rank${Math.abs(remainingRanks) !== 1 ? 's' : ''} over`})</span>}
-                                </div>
-                                <div style={{ display: 'grid', gap: 6 }}>
-                                  {allocs.map((alloc, ai) => {
-                                    const filteredOpts = availableSlOpts.filter((o) => o.value === alloc.id || !usedIds.has(o.value));
-                                    return (
-                                      <div key={ai} style={{ display: 'flex', gap: 6, alignItems: 'flex-end' }}>
-                                        <div style={{ flex: 1 }}>
-                                          <RichSelect
-                                            label={`Spell List ${ai + 1}`}
-                                            value={alloc.id}
-                                            onChange={(v) => updateResolution((r) => ({ ...r, spellListChoices: r.spellListChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, id: v })) }))}
-                                            options={filteredOpts}
-                                            placeholderOption="— Select spell list —"
-                                          />
-                                        </div>
-                                        <div style={{ display: 'flex', alignItems: 'center', gap: 4, paddingBottom: 4 }}>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, spellListChoices: r.spellListChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, ranks: a.ranks - 1 })) }))} disabled={alloc.ranks <= 1}>-</button>
-                                          <span style={{ minWidth: 24, textAlign: 'center' }}>{alloc.ranks}</span>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, spellListChoices: r.spellListChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, ranks: a.ranks + 1 })) }))} disabled={remainingRanks <= 0}>+</button>
-                                        </div>
-                                        <button type="button" onClick={() => updateResolution((r) => ({ ...r, spellListChoices: r.spellListChoices.map((g, gIdx) => gIdx !== gi ? g : g.filter((_, aIdx) => aIdx !== ai)) }))} style={{ marginBottom: 4 }}>×</button>
-                                      </div>
-                                    );
-                                  })}
-                                </div>
-                                {allocs.length < choiceDef.numChoices && remainingRanks > 0 && (
-                                  <button type="button" style={{ marginTop: 4 }} onClick={() => updateResolution((r) => ({ ...r, spellListChoices: r.spellListChoices.map((g, gIdx) => gIdx !== gi ? g : [...g, { id: '', ranks: 1 }]) }))}>+ Add spell list</button>
-                                )}
-                              </div>
-                            );
-                          })}
-                        </div>
-                      </div>
-                    )}
-
-                    {(currentTp.spellListCategoryRankChoices ?? []).some((c) => c.numChoices > 0) && (
-                      <div style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                        <strong>Spell List Category Choices</strong>
-                        <div style={{ display: 'grid', gap: 10, marginTop: 6 }}>
-                          {(currentTp.spellListCategoryRankChoices ?? []).map((choiceDef, gi) => {
-                            if (choiceDef.numChoices <= 0) return null;
-                            const allocs = currentResolution.spellListCategoryChoices[gi] ?? [];
-                            const totalAllocated = allocs.reduce((s, a) => s + a.ranks, 0);
-                            const remainingRanks = choiceDef.value - totalAllocated;
-                            const usedIds = new Set(allocs.map((a) => a.id).filter(Boolean));
-                            const aggregatedSlOpts = choiceDef.options.flatMap((catId) =>
-                              (characterBuilder.categorySpellLists.find((c) => c.category === catId)?.spellLists ?? [])
-                                .map(toSpellListRichOption)
-                            );
-                            return (
-                              <div key={gi}>
-                                <div style={{ color: 'var(--muted)', fontSize: '0.9em', marginBottom: 4 }}>
-                                  Distribute {choiceDef.value} rank{choiceDef.value !== 1 ? 's' : ''} across up to {choiceDef.numChoices} spell list{choiceDef.numChoices !== 1 ? 's' : ''}
-                                  {remainingRanks !== 0 && <span style={{ color: remainingRanks > 0 ? 'var(--warning, #b96c00)' : '#b00020', marginLeft: 8, fontWeight: 600 }}>({remainingRanks > 0 ? `${remainingRanks} rank${remainingRanks !== 1 ? 's' : ''} unallocated` : `${Math.abs(remainingRanks)} rank${Math.abs(remainingRanks) !== 1 ? 's' : ''} over`})</span>}
-                                </div>
-                                <div style={{ display: 'grid', gap: 6 }}>
-                                  {allocs.map((alloc, ai) => {
-                                    const filteredOpts = aggregatedSlOpts.filter((o) => o.value === alloc.id || !usedIds.has(o.value));
-                                    return (
-                                      <div key={ai} style={{ display: 'flex', gap: 6, alignItems: 'flex-end' }}>
-                                        <div style={{ flex: 1 }}>
-                                          <RichSelect
-                                            label={`Spell List ${ai + 1}`}
-                                            value={alloc.id}
-                                            onChange={(v) => updateResolution((r) => ({ ...r, spellListCategoryChoices: r.spellListCategoryChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, id: v })) }))}
-                                            options={filteredOpts}
-                                            placeholderOption="— Select spell list —"
-                                          />
-                                        </div>
-                                        <div style={{ display: 'flex', alignItems: 'center', gap: 4, paddingBottom: 4 }}>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, spellListCategoryChoices: r.spellListCategoryChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, ranks: a.ranks - 1 })) }))} disabled={alloc.ranks <= 1}>-</button>
-                                          <span style={{ minWidth: 24, textAlign: 'center' }}>{alloc.ranks}</span>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, spellListCategoryChoices: r.spellListCategoryChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, ranks: a.ranks + 1 })) }))} disabled={remainingRanks <= 0}>+</button>
-                                        </div>
-                                        <button type="button" onClick={() => updateResolution((r) => ({ ...r, spellListCategoryChoices: r.spellListCategoryChoices.map((g, gIdx) => gIdx !== gi ? g : g.filter((_, aIdx) => aIdx !== ai)) }))} style={{ marginBottom: 4 }}>×</button>
-                                      </div>
-                                    );
-                                  })}
-                                </div>
-                                {allocs.length < choiceDef.numChoices && remainingRanks > 0 && (
-                                  <button type="button" style={{ marginTop: 4 }} onClick={() => updateResolution((r) => ({ ...r, spellListCategoryChoices: r.spellListCategoryChoices.map((g, gIdx) => gIdx !== gi ? g : [...g, { id: '', ranks: 1 }]) }))}>+ Add spell list</button>
-                                )}
-                              </div>
-                            );
-                          })}
-                        </div>
-                      </div>
-                    )}
-
-                    {(currentTp.lifestyleCategorySkillChoices ?? []).some((c) => c.numChoices > 0) && (
-                      <div style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                        <strong>Lifestyle Skill Choices</strong>
-                        <div style={{ display: 'grid', gap: 10, marginTop: 6 }}>
-                          {(currentTp.lifestyleCategorySkillChoices ?? []).map((choiceDef, gi) => {
-                            if (choiceDef.numChoices <= 0) return null;
-                            const chosen = currentResolution.lifestyleCategorySkillChoices[gi] ?? [];
-                            const poolSkills = choiceDef.options.flatMap((catId) => skillIdsByCategory.get(catId) ?? []);
-                            return (
-                              <div key={gi}>
-                                <div style={{ color: 'var(--muted)', fontSize: '0.9em', marginBottom: 4 }}>
-                                  Choose {choiceDef.numChoices} skill{choiceDef.numChoices !== 1 ? 's' : ''} to gain Lifestyle development — these will be available at Lifestyle cost during extra options.
-                                </div>
-                                <div style={{ display: 'grid', gap: 4 }}>
-                                  {Array.from({ length: choiceDef.numChoices }, (_, si) => {
-                                    const val = chosen[si] ?? '';
-                                    const usedOthers = new Set(chosen.filter((s, i) => i !== si && s));
-                                    const opts = poolSkills.filter((s) => s.id === val || !usedOthers.has(s.id)).map(toSkillRichOption);
-                                    return (
-                                      <RichSelect
-                                        key={si}
-                                        label={`Skill ${si + 1}`}
-                                        value={val}
-                                        onChange={(v) => updateResolution((r) => ({
-                                          ...r,
-                                          lifestyleCategorySkillChoices: r.lifestyleCategorySkillChoices.map((g, gIdx) =>
-                                            gIdx !== gi ? g : g.map((s, sIdx) => sIdx !== si ? s : v),
-                                          ),
-                                        }))}
-                                        options={opts}
-                                        placeholderOption="— Select skill —"
-                                      />
-                                    );
-                                  })}
-                                </div>
-                              </div>
-                            );
-                          })}
-                        </div>
-                      </div>
-                    )}
-
-                    {(currentTp.languageChoices ?? []).some((c) => c.numChoices > 0) && (
-                      <div style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                        <strong>Language Choices</strong>
-                        <div style={{ display: 'grid', gap: 10, marginTop: 6 }}>
-                          {(currentTp.languageChoices ?? []).map((choiceDef, gi) => {
-                            if (choiceDef.numChoices <= 0) return null;
-                            const allocs = currentResolution.languageChoices[gi] ?? [];
-                            const totalAllocated = allocs.reduce((s, a) => s + a.spoken + a.written + a.somatic, 0);
-                            const remainingRanks = choiceDef.value - totalAllocated;
-                            const usedLangIds = new Set(allocs.map((a) => a.languageId).filter(Boolean));
-                            const availableLangOpts = choiceDef.options
-                              .filter((lId) => lId)
-                              .map((lId) => ({ value: lId, label: languageNameById.get(lId) ?? lId }));
-                            return (
-                              <div key={gi}>
-                                <div style={{ color: 'var(--muted)', fontSize: '0.9em', marginBottom: 4 }}>
-                                  Distribute {choiceDef.value} rank{choiceDef.value !== 1 ? 's' : ''} across up to {choiceDef.numChoices} language{choiceDef.numChoices !== 1 ? 's' : ''}
-                                  {remainingRanks !== 0 && <span style={{ color: remainingRanks > 0 ? 'var(--warning, #b96c00)' : '#b00020', marginLeft: 8, fontWeight: 600 }}>({remainingRanks > 0 ? `${remainingRanks} rank${remainingRanks !== 1 ? 's' : ''} unallocated` : `${Math.abs(remainingRanks)} rank${Math.abs(remainingRanks) !== 1 ? 's' : ''} over`})</span>}
-                                </div>
-                                <div style={{ display: 'grid', gap: 8 }}>
-                                  {allocs.map((alloc, ai) => {
-                                    const lang = languageById.get(alloc.languageId);
-                                    const filteredOpts = availableLangOpts.filter((o) => o.value === alloc.languageId || !usedLangIds.has(o.value));
-                                    return (
-                                      <div key={ai} style={{ border: '1px solid var(--border)', borderRadius: 4, padding: 6 }}>
-                                        <div style={{ display: 'flex', gap: 6, alignItems: 'flex-end', marginBottom: 6 }}>
-                                          <div style={{ flex: 1 }}>
-                                            <LabeledSelect
-                                              label={`Language ${ai + 1}`}
-                                              value={alloc.languageId}
-                                              onChange={(v) => updateResolution((r) => ({
-                                                ...r,
-                                                languageChoices: r.languageChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { languageId: v, spoken: 0, written: 0, somatic: 0 })),
-                                              }))}
-                                              options={filteredOpts}
-                                              placeholderOption="— Select language —"
-                                            />
-                                          </div>
-                                          <button type="button" onClick={() => updateResolution((r) => ({ ...r, languageChoices: r.languageChoices.map((g, gIdx) => gIdx !== gi ? g : g.filter((_, aIdx) => aIdx !== ai)) }))} style={{ marginBottom: 4 }}>×</button>
-                                        </div>
-                                        {alloc.languageId && (
-                                          <div style={{ display: 'flex', gap: 12, flexWrap: 'wrap' }}>
-                                            {(['spoken', 'written', 'somatic'] as const).map((ability) => {
-                                              const enabled = ability === 'spoken' ? (lang?.isSpoken ?? false) : ability === 'written' ? (lang?.isWritten ?? false) : (lang?.isSomatic ?? false);
-                                              if (!enabled) return null;
-                                              const val = alloc[ability];
-                                              return (
-                                                <div key={ability} style={{ display: 'flex', alignItems: 'center', gap: 4 }}>
-                                                  <span style={{ textTransform: 'capitalize', minWidth: 52 }}>{ability}</span>
-                                                  <button type="button" onClick={() => updateResolution((r) => ({ ...r, languageChoices: r.languageChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, [ability]: a[ability] - 1 })) }))} disabled={val <= 0}>-</button>
-                                                  <span style={{ minWidth: 24, textAlign: 'center' }}>{val}</span>
-                                                  <button type="button" onClick={() => updateResolution((r) => ({ ...r, languageChoices: r.languageChoices.map((g, gIdx) => gIdx !== gi ? g : g.map((a, aIdx) => aIdx !== ai ? a : { ...a, [ability]: a[ability] + 1 })) }))} disabled={remainingRanks <= 0}>+</button>
-                                                </div>
-                                              );
-                                            })}
-                                          </div>
-                                        )}
-                                      </div>
-                                    );
-                                  })}
-                                </div>
-                                {allocs.length < choiceDef.numChoices && (
-                                  <button type="button" style={{ marginTop: 4 }} onClick={() => updateResolution((r) => ({ ...r, languageChoices: r.languageChoices.map((g, gIdx) => gIdx !== gi ? g : [...g, { languageId: '', spoken: 0, written: 0, somatic: 0 }]) }))}>+ Add language</button>
-                                )}
-                              </div>
-                            );
-                          })}
-                        </div>
-                      </div>
-                    )}
-
-                    {resError && <div style={{ color: '#b00020' }}>{resError}</div>}
-
-                    <div style={{ display: 'flex', gap: 8 }}>
-                      <button
-                        type="button"
-                        onClick={() => {
-                          if (levellingResolvingTpIndex === 0) {
-                            setLevellingSubstep('selecting');
-                          } else {
-                            setLevellingResolvingTpIndex(levellingResolvingTpIndex - 1);
-                          }
-                        }}
-                      >
-                        ← {levellingResolvingTpIndex === 0 ? 'TP Selection' : 'Previous TP'}
-                      </button>
-                      <button
-                        type="button"
-                        disabled={!!resError}
-                        onClick={() => {
-                          if (levellingResolvingTpIndex + 1 >= tpsRequiringResolution.length) {
-                            setLevellingSubstep('purchasing');
-                          } else {
-                            setLevellingResolvingTpIndex(levellingResolvingTpIndex + 1);
-                          }
-                        }}
-                      >
-                        {levellingResolvingTpIndex + 1 >= tpsRequiringResolution.length ? 'Finish Resolving →' : 'Next TP →'}
-                      </button>
-                    </div>
-                  </div>
-                );
-              })()}
-
-
-              {/* PURCHASING sub-step */}
-              {levellingSubstep === 'purchasing' && (
-                <>
-                  {selectedLevellingTrainingPackages.length > 0 && (
-                    <div style={{ border: '1px solid var(--border)', borderRadius: 8, padding: 10 }}>
-                      <h4 style={{ margin: '0 0 4px' }}>Selected Training Packages</h4>
-                      {selectedLevellingTrainingPackages.map((tp) => (
-                        <div key={tp.id} style={{ color: 'var(--muted)' }}>
-                          {tp.name}{tp.lifestyle ? ' (Lifestyle)' : ''} — {tpCostMap.get(tp.id) ?? 0} DP
-                        </div>
-                      ))}
-                    </div>
-                  )}
-
-                  {/* Stat Gain Rolls */}
-                  <div style={{ border: '1px solid var(--border)', borderRadius: 8, padding: 10 }}>
-                    <h4 style={{ margin: '0 0 8px' }}>Stat Gain Rolls ({STAT_GAIN_DP_COST} DP each)</h4>
-                    {levellingStatGains.length > 0 && (
-                      <div style={{ display: 'grid', gap: 6, marginBottom: 8 }}>
-                        {levellingStatGains.map((stat, i) => {
-                          const availableOptions = STATS
-                            .filter((s) => !levellingStatGainsUnavailable.has(s) && (s === stat || !levellingStatGains.includes(s)))
-                            .map((s) => ({ value: s, label: s }));
-                          return (
-                            <div key={i} style={{ display: 'flex', gap: 8, alignItems: 'center' }}>
-                              <LabeledSelect
-                                label={`Stat Gain #${i + 1}`}
-                                value={stat}
-                                onChange={(v) => {
-                                  setLevellingStatGains((prev) => prev.map((s, idx) => (idx === i ? v as Stat : s)));
-                                }}
-                                options={availableOptions}
-                              />
-                              <button
-                                type="button"
-                                onClick={() => setLevellingStatGains((prev) => prev.filter((_, idx) => idx !== i))}
-                              >
-                                Remove
-                              </button>
-                            </div>
-                          );
-                        })}
-                      </div>
-                    )}
-                    <button
-                      type="button"
-                      onClick={() => {
-                        const nextStat = STATS.find((s) => !levellingStatGainsUnavailable.has(s) && !levellingStatGains.includes(s));
-                        if (nextStat) setLevellingStatGains((prev) => [...prev, nextStat]);
-                      }}
-                      disabled={
-                        levellingDpRemaining < STAT_GAIN_DP_COST
-                        || STATS.every((s) => levellingStatGainsUnavailable.has(s) || levellingStatGains.includes(s))
-                      }
-                    >
-                      Add Stat Gain Roll
-                    </button>
-                  </div>
-
-                  {/* Skill Rank Purchases */}
-                  <div style={{ border: '1px solid var(--border)', borderRadius: 8, padding: 10 }}>
-                    <h4 style={{ margin: '0 0 8px' }}>Skill Ranks</h4>
-                    {levellingSkillPurchases.length > 0 && (
-                      <div style={{ display: 'grid', gap: 8, marginBottom: 8 }}>
-                        {levellingSkillPurchases.map((purchase, i) => {
-                          const skillName = skillNameById.get(purchase.id) ?? purchase.id;
-                          const categoryId = skillCategoryMap.get(purchase.id) ?? '';
-                          const costElements = categoryCostMap.get(categoryId) ?? [];
-                          const devType = skillDevTypeMap.get(purchase.id);
-                          const tpRanksForSkill = tpGrantedSkillRankCounts.get(purchase.id) ?? 0;
-                          const maxPurch = getSkillMaxDpPurchases(costElements, devType, tpRanksForSkill);
-                          const ranksPerPurchase = getSkillRanksPerPurchase(devType);
-                          const totalRanks = devType === 'Restricted' ? purchase.purchases : purchase.purchases * ranksPerPurchase;
-                          const totalCost = getSkillDpCostWithTpOffset(costElements, devType, purchase.purchases, tpRanksForSkill);
-                          const nextPurchaseCost = purchase.purchases < maxPurch
-                            ? getSkillDpCostWithTpOffset(costElements, devType, purchase.purchases + 1, tpRanksForSkill) - totalCost
-                            : 0;
-                          const isWeaponGroupSkill = weaponGroupSkillIds.has(purchase.id);
-                          const needsSubcategory = mandatorySubcategorySkillIds.has(purchase.id);
-                          const existingSkillRanks = preLevellingRanks.skillRanks
-                            .filter((r) => r.id === purchase.id && (purchase.subcategory ? r.subcategory === purchase.subcategory : true))
-                            .reduce((sum, r) => sum + r.value, 0);
-                          const afterSkillRanks = existingSkillRanks + tpRanksForSkill + totalRanks;
-
-                          return (
-                            <div key={purchase.id} style={{ border: '1px solid var(--border)', borderRadius: 6, padding: 8 }}>
-                              <div style={{ display: 'grid', gridTemplateColumns: '1fr auto auto', alignItems: 'center', gap: 8 }}>
-                                <div style={{ display: 'flex', alignItems: 'center', gap: 8, flexWrap: 'wrap' }}>
-                                  <div>
-                                    <strong>{skillName}</strong>
-                                    <span style={{ color: 'var(--muted)', marginLeft: 6 }}>({categoryGroupNameById.get(categoryId) ?? categoryId})</span>
-                                    {devType && <span style={{ color: 'var(--muted)', marginLeft: 6 }}>[{devType}]</span>}
-                                    <span style={{ color: 'var(--muted)', marginLeft: 6 }}>
-                                      — {existingSkillRanks} → {afterSkillRanks} rank{afterSkillRanks !== 1 ? 's' : ''}, {totalCost} DP
-                                    </span>
-                                  </div>
-                                  {needsSubcategory && (
-                                    isWeaponGroupSkill ? (
-                                      <LabeledSelect
-                                        label="Weapon type"
-                                        hideLabel={true}
-                                        value={purchase.subcategory}
-                                        onChange={(v) => setLevellingSkillPurchases((prev) =>
-                                          prev.map((p, idx) => idx === i ? { ...p, subcategory: v } : p),
-                                        )}
-                                        options={weaponTypeOptionsBySkillId.get(purchase.id) ?? []}
-                                        placeholderOption="— Select weapon type —"
-                                      />
-                                    ) : (
-                                      <LabeledInput
-                                        label="Subcategory"
-                                        hideLabel={true}
-                                        value={purchase.subcategory}
-                                        onChange={(v) => setLevellingSkillPurchases((prev) =>
-                                          prev.map((p, idx) => idx === i ? { ...p, subcategory: v } : p),
-                                        )}
-                                        placeholder="Subcategory"
-                                      />
-                                    )
-                                  )}
-                                </div>
-                                <div style={{ display: 'flex', alignItems: 'center', gap: 6 }}>
-                                  <button
-                                    type="button"
-                                    disabled={purchase.purchases <= 0}
-                                    onClick={() => setLevellingSkillPurchases((prev) =>
-                                      prev.map((p, idx) => idx === i ? { ...p, purchases: p.purchases - 1 } : p),
-                                    )}
-                                  >
-                                    -
-                                  </button>
-                                  <span style={{ minWidth: 20, textAlign: 'center' }}>{purchase.purchases}</span>
-                                  <button
-                                    type="button"
-                                    disabled={purchase.purchases >= maxPurch || nextPurchaseCost > levellingDpRemaining}
-                                    onClick={() => setLevellingSkillPurchases((prev) =>
-                                      prev.map((p, idx) => idx === i ? { ...p, purchases: p.purchases + 1 } : p),
-                                    )}
-                                  >
-                                    +
-                                  </button>
-                                </div>
-                                <button
-                                  type="button"
-                                  onClick={() => setLevellingSkillPurchases((prev) => prev.filter((_, idx) => idx !== i))}
-                                >
-                                  Remove
-                                </button>
-                              </div>
-                            </div>
-                          );
-                        })}
-                      </div>
-                    )}
-                    <div style={{ display: 'flex', flexDirection: 'column', gap: 8 }}>
-                      <div style={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: 8 }}>
-                        <RichSelect
-                          label="Skill Category"
-                          hideLabel={true}
-                          value={levellingSkillCategoryFilter}
-                          onChange={(v) => {
-                            setLevellingSkillCategoryFilter(v);
-                            setLevellingSkillPendingId('');
-                            setLevellingSkillPendingSubcategory('');
-                          }}
-                          options={levellingSkillCategoryOptions}
-                          placeholderOption="— Filter by category —"
-                        />
-                        <RichSelect
-                          label="Add Skill"
-                          hideLabel={true}
-                          value=""
-                          onChange={(v) => {
-                            if (!v) return;
-                            if (mandatorySubcategorySkillIds.has(v)) {
-                              setLevellingSkillPendingId(v);
-                              setLevellingSkillPendingSubcategory('');
-                            } else {
-                              setLevellingSkillPurchases((prev) => [...prev, { id: v, subcategory: '', purchases: 1 }]);
-                              setLevellingSkillPendingId('');
-                              setLevellingSkillPendingSubcategory('');
-                            }
-                          }}
-                          options={levellingSkillOptions}
-                          placeholderOption={levellingSkillCategoryFilter ? '— Select skill —' : '— Select a category first —'}
-                        />
-                      </div>
-                      {levellingSkillPendingId && (
-                        <div style={{ display: 'flex', gap: 8, alignItems: 'center' }}>
-                          <span style={{ fontWeight: 500 }}>{skillNameById.get(levellingSkillPendingId) ?? levellingSkillPendingId}:</span>
-                          {weaponGroupSkillIds.has(levellingSkillPendingId) ? (
-                            <LabeledSelect
-                              label="Weapon type"
-                              hideLabel={true}
-                              value={levellingSkillPendingSubcategory}
-                              onChange={(v) => setLevellingSkillPendingSubcategory(v)}
-                              options={weaponTypeOptionsBySkillId.get(levellingSkillPendingId) ?? []}
-                              placeholderOption="— Select weapon type —"
-                            />
-                          ) : (
-                            <LabeledInput
-                              label="Subcategory"
-                              hideLabel={true}
-                              value={levellingSkillPendingSubcategory}
-                              onChange={(v) => setLevellingSkillPendingSubcategory(v)}
-                              placeholder="Subcategory"
-                            />
-                          )}
-                          <button
-                            type="button"
-                            disabled={!levellingSkillPendingSubcategory.trim()}
-                            onClick={() => {
-                              setLevellingSkillPurchases((prev) => [...prev, { id: levellingSkillPendingId, subcategory: levellingSkillPendingSubcategory.trim(), purchases: 1 }]);
-                              setLevellingSkillPendingId('');
-                              setLevellingSkillPendingSubcategory('');
-                            }}
-                          >
-                            Add
-                          </button>
-                          <button
-                            type="button"
-                            onClick={() => {
-                              setLevellingSkillPendingId('');
-                              setLevellingSkillPendingSubcategory('');
-                            }}
-                          >
-                            Cancel
-                          </button>
-                        </div>
-                      )}
-                    </div>
-                  </div>
-
-                  {/* Category Rank Purchases */}
-                  <div style={{ border: '1px solid var(--border)', borderRadius: 8, padding: 10 }}>
-                    <h4 style={{ margin: '0 0 8px' }}>Skill Category Ranks</h4>
-                    {levellingCategoryPurchases.length > 0 && (
-                      <div style={{ display: 'grid', gap: 6, marginBottom: 8 }}>
-                        {levellingCategoryPurchases.map((purchase, i) => {
-                          const catName = categoryNameById.get(purchase.id) ?? purchase.id;
-                          const costElements = categoryCostMap.get(purchase.id) ?? [];
-                          const tpRanksForCat = tpGrantedCategoryRankCounts.get(purchase.id) ?? 0;
-                          const maxPurch = getCategoryMaxDpPurchases(costElements, tpRanksForCat);
-                          const totalCost = getCategoryDpCostWithTpOffset(costElements, purchase.purchases, tpRanksForCat);
-                          const nextCost = purchase.purchases < maxPurch
-                            ? getCategoryDpCostWithTpOffset(costElements, purchase.purchases + 1, tpRanksForCat) - totalCost
-                            : 0;
-                          const existingCatRanks = preLevellingRanks.categoryRanks
-                            .filter((r) => r.id === purchase.id)
-                            .reduce((sum, r) => sum + r.value, 0);
-                          const afterCatRanks = existingCatRanks + tpRanksForCat + purchase.purchases;
-
-                          return (
-                            <div key={purchase.id} style={{ display: 'grid', gridTemplateColumns: '1fr auto auto', alignItems: 'center', gap: 8 }}>
-                              <span>
-                                {catName}
-                                <span style={{ color: 'var(--muted)', marginLeft: 6 }}>
-                                  — {existingCatRanks} → {afterCatRanks} rank{afterCatRanks !== 1 ? 's' : ''}, {totalCost} DP
-                                </span>
-                              </span>
-                              <div style={{ display: 'flex', alignItems: 'center', gap: 6 }}>
-                                <button
-                                  type="button"
-                                  disabled={purchase.purchases <= 0}
-                                  onClick={() => setLevellingCategoryPurchases((prev) =>
-                                    prev.map((p, idx) => idx === i ? { ...p, purchases: p.purchases - 1 } : p),
-                                  )}
-                                >
-                                  -
-                                </button>
-                                <span style={{ minWidth: 20, textAlign: 'center' }}>{purchase.purchases}</span>
-                                <button
-                                  type="button"
-                                  disabled={purchase.purchases >= maxPurch || nextCost > levellingDpRemaining}
-                                  onClick={() => setLevellingCategoryPurchases((prev) =>
-                                    prev.map((p, idx) => idx === i ? { ...p, purchases: p.purchases + 1 } : p),
-                                  )}
-                                >
-                                  +
-                                </button>
-                              </div>
-                              <button
-                                type="button"
-                                onClick={() => setLevellingCategoryPurchases((prev) => prev.filter((_, idx) => idx !== i))}
-                              >
-                                Remove
-                              </button>
-                            </div>
-                          );
-                        })}
-                      </div>
-                    )}
-                    <RichSelect
-                      label="Add Category"
-                      value=""
-                      hideLabel={true}
-                      onChange={(v) => {
-                        if (v) {
-                          setLevellingCategoryPurchases((prev) => [...prev, { id: v, purchases: 1 }]);
-                        }
-                      }}
-                      options={levellingCategoryOptions}
-                      placeholderOption="— Add category —"
-                    />
-                  </div>
-
-                  {/* Spell List Rank Purchases */}
-                  <div style={{ border: '1px solid var(--border)', borderRadius: 8, padding: 10 }}>
-                    <h4 style={{ margin: '0 0 8px' }}>Spell List Ranks</h4>
-                    {levellingSpellListPurchases.length > 0 && (
-                      <div style={{ display: 'grid', gap: 6, marginBottom: 8 }}>
-                        {levellingSpellListPurchases.map((purchase, i) => {
-                          const slName = spellListNameById.get(purchase.id) ?? purchase.id;
-                          const catEntry = characterBuilder.categorySpellLists.find((c) => c.spellLists.includes(purchase.id));
-                          const costElements = catEntry ? (categoryCostMap.get(catEntry.category) ?? []) : [];
-                          const tpRanksForSl = tpGrantedSpellListRankCounts.get(purchase.id) ?? 0;
-                          const maxPurch = getCategoryMaxDpPurchases(costElements, tpRanksForSl);
-                          const totalCost = getCategoryDpCostWithTpOffset(costElements, purchase.purchases, tpRanksForSl);
-                          const nextCost = purchase.purchases < maxPurch
-                            ? getCategoryDpCostWithTpOffset(costElements, purchase.purchases + 1, tpRanksForSl) - totalCost
-                            : 0;
-                          const existingSlRanks = preLevellingRanks.spellListRanks
-                            .filter((r) => r.id === purchase.id)
-                            .reduce((sum, r) => sum + r.value, 0);
-                          const afterSlRanks = existingSlRanks + tpRanksForSl + purchase.purchases;
-
-                          return (
-                            <div key={purchase.id} style={{ display: 'grid', gridTemplateColumns: '1fr auto auto', alignItems: 'center', gap: 8 }}>
-                              <span>
-                                {slName}
-                                <span style={{ color: 'var(--muted)', marginLeft: 6 }}>
-                                  — {existingSlRanks} → {afterSlRanks} rank{afterSlRanks !== 1 ? 's' : ''}, {totalCost} DP
-                                </span>
-                              </span>
-                              <div style={{ display: 'flex', alignItems: 'center', gap: 6 }}>
-                                <button
-                                  type="button"
-                                  disabled={purchase.purchases <= 0}
-                                  onClick={() => setLevellingSpellListPurchases((prev) =>
-                                    prev.map((p, idx) => idx === i ? { ...p, purchases: p.purchases - 1 } : p),
-                                  )}
-                                >
-                                  -
-                                </button>
-                                <span style={{ minWidth: 20, textAlign: 'center' }}>{purchase.purchases}</span>
-                                <button
-                                  type="button"
-                                  disabled={purchase.purchases >= maxPurch || nextCost > levellingDpRemaining}
-                                  onClick={() => setLevellingSpellListPurchases((prev) =>
-                                    prev.map((p, idx) => idx === i ? { ...p, purchases: p.purchases + 1 } : p),
-                                  )}
-                                >
-                                  +
-                                </button>
-                              </div>
-                              <button
-                                type="button"
-                                onClick={() => setLevellingSpellListPurchases((prev) => prev.filter((_, idx) => idx !== i))}
-                              >
-                                Remove
-                              </button>
-                            </div>
-                          );
-                        })}
-                      </div>
-                    )}
-                    {levellingAddingSpellList ? (
-                      <div style={{ display: 'grid', gap: 8 }}>
-                        <RichSelect
-                          label="Spell Category"
-                          hideLabel={true}
-                          value={levellingSelectedSpellCategory}
-                          onChange={(v) => setLevellingSelectedSpellCategory(v)}
-                          options={levellingSpellCategoryOptions}
-                          placeholderOption="— Select category —"
-                        />
-                        {levellingSelectedSpellCategory && levellingSpellListsInSelectedCategory.length > 0 && (
-                          <RichSelect
-                            label="Spell List"
-                            hideLabel={true}
-                            value=""
-                            onChange={(v) => {
-                              if (v) {
-                                setLevellingSpellListPurchases((prev) => [...prev, { id: v, purchases: 1 }]);
-                                setLevellingSelectedSpellCategory('');
-                                setLevellingAddingSpellList(false);
-                              }
-                            }}
-                            options={levellingSpellListsInSelectedCategory.map((sl) => toSpellListRichOption(sl.id))}
-                            placeholderOption="— Select spell list —"
-                          />
-                        )}
-                        {levellingSelectedSpellCategory && levellingSpellListsInSelectedCategory.length === 0 && (
-                          <div style={{ color: 'var(--muted)' }}>No more spell lists available in this category.</div>
-                        )}
-                        <div>
-                          <button
-                            type="button"
-                            onClick={() => {
-                              setLevellingAddingSpellList(false);
-                              setLevellingSelectedSpellCategory('');
-                            }}
-                          >
-                            Cancel
-                          </button>
-                        </div>
-                      </div>
-                    ) : (
-                      <button
-                        type="button"
-                        disabled={
-                          levellingDpRemaining <= 0
-                          || characterBuilder.categorySpellLists.every((c) =>
-                            c.spellLists.every((slId) => levellingSpellListPurchases.some((p) => p.id === slId)),
-                          )
-                        }
-                        onClick={() => setLevellingAddingSpellList(true)}
-                      >
-                        Add Spell List
-                      </button>
-                    )}
-                  </div>
-
-                  <div style={{ display: 'flex', gap: 8 }}>
-                    <button
-                      type="button"
-                      onClick={() => {
-                        if (tpsRequiringResolution.length === 0) {
-                          setLevellingSubstep('selecting');
-                        } else {
-                          setLevellingResolvingTpIndex(tpsRequiringResolution.length - 1);
-                          setLevellingSubstep('resolving');
-                        }
-                      }}
-                    >
-                      ← Back
-                    </button>
-                    <button
-                      type="button"
-                      onClick={async () => { if (await submitLevelUpgrade()) setStep('summary'); }}
-                      disabled={applying || Boolean(errors.apprenticeship)}
-                    >
-                      {applying ? 'Applying…' : 'Complete Apprenticeship'}
-                    </button>
-                  </div>
-                </>
-              )}
-
-              {errors.apprenticeship && <div style={{ color: '#b00020' }}>{errors.apprenticeship}</div>}
-            </section>
-          )}
-
           {step === 'summary' && (
             <section style={{ display: 'grid', gap: 12 }}>
               <div style={{ border: '1px solid var(--border)', borderRadius: 8, padding: 10 }}>
@@ -5846,8 +3824,6 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
                   <li>Profession: {profession?.name ?? professionId}</li>
                   <li>Realms: {selectedRealms.join(', ') || 'None'}</li>
                   <li>Prime Stats: {primeStats.join(', ') || 'None'}</li>
-                  <li>Training packages: {selectedLevellingTrainingPackages.length > 0 ? selectedLevellingTrainingPackages.map((tp) => tp.name).join(', ') : 'None'}</li>
-                  <li>DP spent: {levellingTotalDpSpent} / {characterBuilder.developmentPoints}</li>
                 </ul>
               </div>
 
@@ -5864,12 +3840,12 @@ export default function CharacterCreationView({ onFinish }: { onFinish?: () => v
             </section>
           )}
 
-          {step !== 'summary' && (
+          {step !== 'summary' && (<>
             <div style={{ display: 'flex', gap: 8 }}>
               <button type="button" onClick={goPrev} disabled={step === 'primary'}>Back</button>
               <button type="button" onClick={goNext} disabled={!canGoNext || savingPrimaryDefinition || savingInitialChoices || savingStats || savingPhysique || savingHobbyChoices || savingBackgroundChoices}>Next</button>
             </div>
-          )}
+          </>)}
         </div>
       </div>
     </>
